RF: Move Color classes to psychopy.colors
