+++ psychopy/colors.py
@@ -5,8 +5,883 @@ from __future__ import absolute_import, print_function
 
 from past.builtins import basestring
 from psychopy.visual import basevisual as vis
+from past.builtins import basestring
+from psychopy.tools.colorspacetools import dkl2rgb, lms2rgb, hsv2rgb  # pylint: disable=W0611
+from psychopy import logging
+import re
+import numpy
+from math import floor, fsum
+
+from psychopy import logging
+
+# Dict of examples of Psychopy Red at 12% opacity in different formats
+colorExamples = {
+    'invis': None,
+    'named': 'crimson',
+    'hex': '#F2545B',
+    'hexa': '#F2545B1E',
+    'rgb': (0.89, -0.35, -0.28),
+    'rgba': (0.89, -0.35, -0.28, -0.76),
+    'rgb1': (0.95, 0.32, 0.36),
+    'rgba1': (0.95, 0.32, 0.36, 0.12),
+    'rgb255': (242, 84, 91),
+    'rgba255': (242, 84, 91, 30),
+    'hsv': (357, 0.65, 0.95),
+    'hsva': (357, 0.65, 0.95, 0.12),
+}
+# Dict of named colours
+colorNames = {
+        "none": (0, 0, 0, -1),
+        "aliceblue": (0.882352941176471, 0.945098039215686, 1, 1.0),
+        "antiquewhite": (0.96078431372549, 0.843137254901961, 0.686274509803922, 1.0),
+        "aqua": (-1, 1, 1, 1.0),
+        "aquamarine": (-0.00392156862745097, 1, 0.662745098039216, 1.0),
+        "azure": (0.882352941176471, 1, 1, 1.0),
+        "beige": (0.92156862745098, 0.92156862745098, 0.725490196078431, 1.0),
+        "bisque": (1, 0.788235294117647, 0.537254901960784, 1.0),
+        "black": (-1, -1, -1, 1.0),
+        "blanchedalmond": (1, 0.843137254901961, 0.607843137254902, 1.0),
+        "blue": (-1, -1, 1, 1.0),
+        "blueviolet": (0.0823529411764705, -0.662745098039216, 0.772549019607843, 1.0),
+        "brown": (0.294117647058824, -0.670588235294118, -0.670588235294118, 1.0),
+        "burlywood": (0.741176470588235, 0.443137254901961, 0.0588235294117647, 1.0),
+        "cadetblue": (-0.254901960784314, 0.23921568627451, 0.254901960784314, 1.0),
+        "chartreuse": (-0.00392156862745097, 1, -1, 1.0),
+        "chocolate": (0.647058823529412, -0.176470588235294, -0.764705882352941, 1.0),
+        "coral": (1, -0.00392156862745097, -0.372549019607843, 1.0),
+        "cornflowerblue": (-0.215686274509804, 0.168627450980392, 0.858823529411765, 1.0),
+        "cornsilk": (1, 0.945098039215686, 0.725490196078431, 1.0),
+        "crimson": (0.725490196078431, -0.843137254901961, -0.529411764705882, 1.0),
+        "cyan": (-1, 1, 1, 1.0),
+        "darkblue": (-1, -1, 0.0901960784313725, 1.0),
+        "darkcyan": (-1, 0.0901960784313725, 0.0901960784313725, 1.0),
+        "darkgoldenrod": (0.443137254901961, 0.0509803921568628, -0.913725490196078, 1.0),
+        "darkgray": (0.325490196078431, 0.325490196078431, 0.325490196078431, 1.0),
+        "darkgreen": (-1, -0.215686274509804, -1, 1.0),
+        "darkgrey": (0.325490196078431, 0.325490196078431, 0.325490196078431, 1.0),
+        "darkkhaki": (0.482352941176471, 0.435294117647059, -0.16078431372549, 1.0),
+        "darkmagenta": (0.0901960784313725, -1, 0.0901960784313725, 1.0),
+        "darkolivegreen": (-0.333333333333333, -0.16078431372549, -0.631372549019608, 1.0),
+        "darkorange": (1, 0.0980392156862746, -1, 1.0),
+        "darkorchid": (0.2, -0.607843137254902, 0.6, 1.0),
+        "darkred": (0.0901960784313725, -1, -1, 1.0),
+        "darksalmon": (0.827450980392157, 0.176470588235294, -0.0431372549019607, 1.0),
+        "darkseagreen": (0.12156862745098, 0.474509803921569, 0.12156862745098, 1.0),
+        "darkslateblue": (-0.435294117647059, -0.52156862745098, 0.0901960784313725, 1.0),
+        "darkslategray": (-0.631372549019608, -0.380392156862745, -0.380392156862745, 1.0),
+        "darkslategrey": (-0.631372549019608, -0.380392156862745, -0.380392156862745, 1.0),
+        "darkturquoise": (-1, 0.615686274509804, 0.63921568627451, 1.0),
+        "darkviolet": (0.16078431372549, -1, 0.654901960784314, 1.0),
+        "deeppink": (1, -0.843137254901961, 0.152941176470588, 1.0),
+        "deepskyblue": (-1, 0.498039215686275, 1, 1.0),
+        "dimgray": (-0.176470588235294, -0.176470588235294, -0.176470588235294, 1.0),
+        "dimgrey": (-0.176470588235294, -0.176470588235294, -0.176470588235294, 1.0),
+        "dodgerblue": (-0.764705882352941, 0.129411764705882, 1, 1.0),
+        "firebrick": (0.396078431372549, -0.733333333333333, -0.733333333333333, 1.0),
+        "floralwhite": (1, 0.96078431372549, 0.882352941176471, 1.0),
+        "forestgreen": (-0.733333333333333, 0.0901960784313725, -0.733333333333333, 1.0),
+        "fuchsia": (1, -1, 1, 1.0),
+        "gainsboro": (0.725490196078431, 0.725490196078431, 0.725490196078431, 1.0),
+        "ghostwhite": (0.945098039215686, 0.945098039215686, 1, 1.0),
+        "gold": (1, 0.686274509803922, -1, 1.0),
+        "goldenrod": (0.709803921568627, 0.294117647058824, -0.749019607843137, 1.0),
+        "gray": (0.00392156862745097, 0.00392156862745097, 0.00392156862745097, 1.0),
+        "grey": (0.00392156862745097, 0.00392156862745097, 0.00392156862745097, 1.0),
+        "green": (-1, 0.00392156862745097, -1, 1.0),
+        "greenyellow": (0.356862745098039, 1, -0.631372549019608, 1.0),
+        "honeydew": (0.882352941176471, 1, 0.882352941176471, 1.0),
+        "hotpink": (1, -0.176470588235294, 0.411764705882353, 1.0),
+        "indianred": (0.607843137254902, -0.27843137254902, -0.27843137254902, 1.0),
+        "indigo": (-0.411764705882353, -1, 0.0196078431372548, 1.0),
+        "ivory": (1, 1, 0.882352941176471, 1.0),
+        "khaki": (0.882352941176471, 0.803921568627451, 0.0980392156862746, 1.0),
+        "lavender": (0.803921568627451, 0.803921568627451, 0.96078431372549, 1.0),
+        "lavenderblush": (1, 0.882352941176471, 0.92156862745098, 1.0),
+        "lawngreen": (-0.0274509803921569, 0.976470588235294, -1, 1.0),
+        "lemonchiffon": (1, 0.96078431372549, 0.607843137254902, 1.0),
+        "lightblue": (0.356862745098039, 0.694117647058824, 0.803921568627451, 1.0),
+        "lightcoral": (0.882352941176471, 0.00392156862745097, 0.00392156862745097, 1.0),
+        "lightcyan": (0.756862745098039, 1, 1, 1.0),
+        "lightgoldenrodyellow": (0.96078431372549, 0.96078431372549, 0.647058823529412, 1.0),
+        "lightgray": (0.654901960784314, 0.654901960784314, 0.654901960784314, 1.0),
+        "lightgreen": (0.129411764705882, 0.866666666666667, 0.129411764705882, 1.0),
+        "lightgrey": (0.654901960784314, 0.654901960784314, 0.654901960784314, 1.0),
+        "lightpink": (1, 0.427450980392157, 0.513725490196078, 1.0),
+        "lightsalmon": (1, 0.254901960784314, -0.0431372549019607, 1.0),
+        "lightseagreen": (-0.749019607843137, 0.396078431372549, 0.333333333333333, 1.0),
+        "lightskyblue": (0.0588235294117647, 0.615686274509804, 0.96078431372549, 1.0),
+        "lightslategray": (-0.0666666666666667, 0.0666666666666667, 0.2, 1.0),
+        "lightslategrey": (-0.0666666666666667, 0.0666666666666667, 0.2, 1.0),
+        "lightsteelblue": (0.380392156862745, 0.537254901960784, 0.741176470588235, 1.0),
+        "lightyellow": (1, 1, 0.756862745098039, 1.0),
+        "lime": (-1, 1, -1, 1.0),
+        "limegreen": (-0.607843137254902, 0.607843137254902, -0.607843137254902, 1.0),
+        "linen": (0.96078431372549, 0.882352941176471, 0.803921568627451, 1.0),
+        "magenta": (1, -1, 1, 1.0),
+        "maroon": (0.00392156862745097, -1, -1, 1.0),
+        "mediumaquamarine": (-0.2, 0.607843137254902, 0.333333333333333, 1.0),
+        "mediumblue": (-1, -1, 0.607843137254902, 1.0),
+        "mediumorchid": (0.458823529411765, -0.333333333333333, 0.654901960784314, 1.0),
+        "mediumpurple": (0.152941176470588, -0.12156862745098, 0.717647058823529, 1.0),
+        "mediumseagreen": (-0.529411764705882, 0.403921568627451, -0.113725490196078, 1.0),
+        "mediumslateblue": (-0.0352941176470588, -0.184313725490196, 0.866666666666667, 1.0),
+        "mediumspringgreen": (-1, 0.96078431372549, 0.207843137254902, 1.0),
+        "mediumturquoise": (-0.435294117647059, 0.63921568627451, 0.6, 1.0),
+        "mediumvioletred": (0.56078431372549, -0.835294117647059, 0.0431372549019609, 1.0),
+        "midnightblue": (-0.803921568627451, -0.803921568627451, -0.12156862745098, 1.0),
+        "mintcream": (0.92156862745098, 1, 0.96078431372549, 1.0),
+        "mistyrose": (1, 0.788235294117647, 0.764705882352941, 1.0),
+        "moccasin": (1, 0.788235294117647, 0.419607843137255, 1.0),
+        "navajowhite": (1, 0.741176470588235, 0.356862745098039, 1.0),
+        "navy": (-1, -1, 0.00392156862745097, 1.0),
+        "oldlace": (0.984313725490196, 0.92156862745098, 0.803921568627451, 1.0),
+        "olive": (0.00392156862745097, 0.00392156862745097, -1, 1.0),
+        "olivedrab": (-0.16078431372549, 0.113725490196078, -0.725490196078431, 1.0),
+        "orange": (1, 0.294117647058824, -1, 1.0),
+        "orangered": (1, -0.458823529411765, -1, 1.0),
+        "orchid": (0.709803921568627, -0.12156862745098, 0.67843137254902, 1.0),
+        "palegoldenrod": (0.866666666666667, 0.819607843137255, 0.333333333333333, 1.0),
+        "palegreen": (0.192156862745098, 0.968627450980392, 0.192156862745098, 1.0),
+        "paleturquoise": (0.372549019607843, 0.866666666666667, 0.866666666666667, 1.0),
+        "palevioletred": (0.717647058823529, -0.12156862745098, 0.152941176470588, 1.0),
+        "papayawhip": (1, 0.874509803921569, 0.670588235294118, 1.0),
+        "peachpuff": (1, 0.709803921568627, 0.450980392156863, 1.0),
+        "peru": (0.607843137254902, 0.0431372549019609, -0.505882352941176, 1.0),
+        "pink": (1, 0.505882352941176, 0.592156862745098, 1.0),
+        "plum": (0.733333333333333, 0.254901960784314, 0.733333333333333, 1.0),
+        "powderblue": (0.380392156862745, 0.756862745098039, 0.803921568627451, 1.0),
+        "purple": (0.00392156862745097, -1, 0.00392156862745097, 1.0),
+        "red": (1, -1, -1, 1.0),
+        "rosybrown": (0.474509803921569, 0.12156862745098, 0.12156862745098, 1.0),
+        "royalblue": (-0.490196078431373, -0.176470588235294, 0.764705882352941, 1.0),
+        "saddlebrown": (0.0901960784313725, -0.458823529411765, -0.850980392156863, 1.0),
+        "salmon": (0.96078431372549, 0.00392156862745097, -0.105882352941176, 1.0),
+        "sandybrown": (0.913725490196079, 0.286274509803922, -0.247058823529412, 1.0),
+        "seagreen": (-0.63921568627451, 0.0901960784313725, -0.317647058823529, 1.0),
+        "seashell": (1, 0.92156862745098, 0.866666666666667, 1.0),
+        "sienna": (0.254901960784314, -0.356862745098039, -0.647058823529412, 1.0),
+        "silver": (0.505882352941176, 0.505882352941176, 0.505882352941176, 1.0),
+        "skyblue": (0.0588235294117647, 0.615686274509804, 0.843137254901961, 1.0),
+        "slateblue": (-0.168627450980392, -0.294117647058823, 0.607843137254902, 1.0),
+        "slategray": (-0.12156862745098, 0.00392156862745097, 0.129411764705882, 1.0),
+        "slategrey": (-0.12156862745098, 0.00392156862745097, 0.129411764705882, 1.0),
+        "snow": (1, 0.96078431372549, 0.96078431372549, 1.0),
+        "springgreen": (-1, 1, -0.00392156862745097, 1.0),
+        "steelblue": (-0.450980392156863, 0.0196078431372548, 0.411764705882353, 1.0),
+        "tan": (0.647058823529412, 0.411764705882353, 0.0980392156862746, 1.0),
+        "teal": (-1, 0.00392156862745097, 0.00392156862745097, 1.0),
+        "thistle": (0.694117647058824, 0.498039215686275, 0.694117647058824, 1.0),
+        "tomato": (1, -0.223529411764706, -0.443137254901961, 1.0),
+        "turquoise": (-0.498039215686275, 0.756862745098039, 0.631372549019608, 1.0),
+        "violet": (0.866666666666667, 0.0196078431372548, 0.866666666666667, 1.0),
+        "wheat": (0.92156862745098, 0.741176470588235, 0.403921568627451, 1.0),
+        "white": (1, 1, 1, 1.0),
+        "whitesmoke": (0.92156862745098, 0.92156862745098, 0.92156862745098, 1.0),
+        "yellow": (1, 1, -1, 1.0),
+        "yellowgreen": (0.207843137254902, 0.607843137254902, -0.607843137254902, 1.0)
+    }
+# Shorthand for common regexpressions
+_255 = '(\d|\d\d|1\d\d|2[0-4]\d|25[0-5])'
+_360 = '(\d|\d\d|[12]\d\d|3[0-5]\d|360)'
+_100 = '(\d|\d\d|100)'
+_1 = '(0|1|1.0*|0\.\d*)'
+_lbr = '[\[\(]\s*'
+_rbr = '\s*[\]\)]'
+# Dict of regexpressions for different formats
+colorSpaces = {
+    'named': re.compile("|".join(list(colorNames))), # A named colour space
+    'hex': re.compile('#[\dabcdefABCDEF]{6}'), # Hex
+    'hexa': re.compile('#[\dabcdefABCDEF]{8}'), # Hex + alpha
+    'rgb': re.compile(_lbr+'\-?'+_1+',\s*'+'\-?'+_1+',\s*'+'\-?'+_1+_rbr), # RGB from -1 to 1
+    'rgba': re.compile(_lbr+'\-?'+_1+',\s*'+'\-?'+_1+',\s*'+'\-?'+_1+',\s*'+'\-?'+_1+_rbr),  # RGB + alpha from -1 to 1
+    'rgb1': re.compile(_lbr+_1+',\s*'+_1+',\s*'+_1+_rbr),  # RGB from 0 to 1
+    'rgba1': re.compile(_lbr+_1+',\s*'+_1+',\s*'+_1+',\s*'+_1+_rbr),  # RGB + alpha from 0 to 1
+    'rgb255': re.compile(_lbr+_255+',\s*'+_255+',\s*'+_255+_rbr), # RGB from 0 to 255
+    'rgba255': re.compile(_lbr+_255+',\s*'+_255+',\s*'+_255+',\s*'+_255+_rbr), # RGB + alpha from 0 to 255
+    'hsv': re.compile(_lbr+_360+'\°?'+',\s*'+_1+',\s*'+_1+_rbr), # HSV with hue from 0 to 360 and saturation/vibrancy from 0 to 1
+    'hsva': re.compile(_lbr+_360+'\°?'+',\s*'+_1+',\s*'+_1+',\s*'+_1+_rbr), # HSV with hue from 0 to 360 and saturation/vibrancy from 0 to 1 + alpha from 0 to 1
+}
 
 
+class Color(object):
+    """A class to store colour details, knows what colour space it's in and can supply colours in any space"""
+
+    def __init__(self, color=None, space=None):
+        self._cache = {}
+        self.set(color=color, space=space)
+
+    def set(self, color=None, space=None):
+        """Set the colour of this object - essentially the same as what happens on creation, but without
+        having to initialise a new object"""
+        if isinstance(color, numpy.ndarray):
+            color = tuple(float(c) for c in color)
+        if space in ['rgb255', 'rgba255']:
+            color = tuple(int(c) for c in color)
+        # If input is a Color object, duplicate all settings
+        if isinstance(color, Color):
+            self._requested = color._requested
+            self._requestedSpace = color._requestedSpace
+            self.rgba = color.rgba
+            return
+        # Store requested colour and space (or defaults, if none given)
+        self._requested = color if color is not None else None
+        self._requestedSpace = space \
+            if space and space in self.getSpace(self._requested, debug=True) \
+            else self.getSpace(self._requested)
+        if isinstance(self._requestedSpace, (list, type(None))):
+            logging.error("Color space could not be determined by values supplied, please specify a color space.")
+            return
+
+        # Convert to lingua franca
+        if self._requestedSpace:
+            setattr(self, self._requestedSpace, self._requested)
+        else:
+            self.named = None
+
+    def __repr__(self):
+        """If colour is printed, it will display its class and value"""
+        if self.rgba:
+            if self.named:
+                return "<" + self.__class__.__module__ + "." + self.__class__.__name__ + ": " + self.named + ">"
+            else:
+                return "<" + self.__class__.__module__ + "." + self.__class__.__name__ + ": " + str(tuple(round(c,2) for c in self.rgba)) + ">"
+        else:
+            return "<" + self.__class__.__module__ + "." + self.__class__.__name__ + ": " + "Invalid" + ">"
+
+    # ---rich comparisons---
+    def __eq__(self, target):
+        """== will compare RGBA values, rounded to 2dp"""
+        if isinstance(target, Color):
+            return (round(c,2) for c in self.rgba) == (round(c,2) for c in target.rgba)
+        else:
+            return False
+    def __ne__(self, target):
+        """!= will return the opposite of =="""
+        return not self == target
+    def __lt__(self, target):
+        """< will compare brightness"""
+        if isinstance(target, Color):
+            return self.brightness < target.brightness
+        else:
+            return False
+    def __le__(self, target):
+        """<= will compare brightness"""
+        if isinstance(target, Color):
+            return self.brightness <= target.brightness
+        else:
+            return False
+    def __gt__(self, target):
+        """> will compare brightness"""
+        if isinstance(target, Color):
+            return self.brightness > target.brightness
+        else:
+            return False
+    def __ge__(self, target):
+        """>= will compare brightness"""
+        if isinstance(target, Color):
+            return self.brightness >= target.brightness
+        else:
+            return False
+
+    #--operators---
+    def __add__(self, other):
+        if not isinstance(other, Color):
+            if Color.getSpace(other):
+                # Convert to a color if valid
+                other = Color(other)
+            elif AdvancedColor.getSpace(other):
+                # Convert to an advanced color if valid
+                other = AdvancedColor(other)
+            elif isinstance(other, int) or isinstance(other, float):
+                out = self.copy()
+                out.brightness += other
+                return out
+            else:
+                raise ValueError ("unsupported operand type(s) for +: '"
+                                  + self.__class__.__name__ +"' and '"
+                                  + other.__class__.__name__ + "'")
+        if isinstance(other, Color):
+            # If both are colours, average the two and sum their alphas
+            alpha = min(self.alpha + other.alpha, 1)
+            rgb = [None, None, None]
+            selfWeight = self.alpha/(self.alpha+other.alpha)
+            otherWeight = other.alpha/(self.alpha+other.alpha)
+            for c in range(3):
+                rgb[c] = self.rgb1[c]*selfWeight + other.rgb1[c]*otherWeight
+            return Color(rgb+[alpha], 'rgba1')
+
+
+    def copy(self):
+        """Return a duplicate of this colour"""
+        return self.__class__(self._requested, self._requestedSpace)
+
+    @staticmethod
+    def getSpace(color, debug=False):
+        """Find what colour space a colour is from"""
+        if isinstance(color, Color):
+            return color._requestedSpace
+        possible = [space for space in colorSpaces
+                    if colorSpaces[space].fullmatch(str(color).lower())]
+        # Return full list if debug or multiple, else return first value
+        if debug or not len(possible) == 1:
+            return possible
+        else:
+            return possible[0]
+
+    @staticmethod
+    def hue2rgb255(hue):
+        # Work out what segment of the colour wheel we're in
+        seg = floor(hue / 60)
+        seg = seg if seg % 6 else 0
+        # Define values for when a value is decreasing / increasing in a segment
+        _up = (hue % 60) * (255 / 60)
+        _down = 255 - _up
+        _mov = _down if seg%2 else _up # Even segments are down, odd are up
+        # Calculate rgb according to segment
+        if seg == 0:
+            return (255, _mov, 0,)
+        if seg == 1:
+            return (_mov, 255, 0,)
+        if seg == 2:
+            return (0, 255, _mov,)
+        if seg == 3:
+            return (0, _mov, 255,)
+        if seg == 4:
+            return (_mov, 0, 255,)
+        if seg == 5:
+            return (255, 0, _mov,)
+
+    def validate(self, color, against=None, enforce=None):
+        # If not checking against anything, check against everything
+        if not against:
+            against = list(colorSpaces)
+        # If looking for a string, convert from other forms it could be in
+        if enforce == str:
+            color = str(color).lower()
+        # If looking for a tuple, convert from other forms it could be in
+        if enforce == tuple or enforce == (tuple, int):
+            if isinstance(color, str):
+                color = [float(n) for n in color.strip('[]()').split(',')]
+            if isinstance(color, list):
+                color = tuple(color)
+        # If enforcing multiple
+        if enforce == (tuple, int):
+            color = tuple(int(round(c)) for c in color)
+        # Get possible colour spaces
+        possible = Color.getSpace(color)
+        if isinstance(possible, str):
+            possible = [possible]
+        # Return if any matches
+        for space in possible:
+            if space in against:
+                return color
+        # If no matches...
+        self.named = None
+        return None
+
+    # ---adjusters---
+    @property
+    def contrast(self):
+        '''Distance from mid grey in rgb (-1 to 1)'''
+        return abs(sum(self.rgb)/3)
+    @contrast.setter
+    def contrast(self, value):
+        if self.contrast:
+            norm = tuple(c/self.contrast for c in self.rgb)
+        else:
+            norm = self.rgb
+        self.rgba = tuple(max(min(c*value,1),-1) for c in norm) + (self.rgba[-1],)
+
+    @property
+    def saturation(self):
+        '''Saturation in hsv (0 to 1)'''
+        return self.hsv[1]
+    @saturation.setter
+    def saturation(self, value):
+        h,s,v,a = self.hsva
+        s += value
+        s = min(s, 1)
+        s = max(s, 0)
+        self.hsva = (h, s, v, a)
+
+    @property
+    def brightness(self):
+        return sum(self.rgb1)/3
+    @brightness.setter
+    def brightness(self, value):
+        adj = value-self.brightness
+        self.rgba1 = tuple(
+            max(min(c+adj, 1),0)
+            for c in self.rgb1
+        ) + (self.alpha,)
+
+    @property
+    def alpha(self):
+        return self.rgba1[-1]
+    @alpha.setter
+    def alpha(self, value):
+        value = min(value,1)
+        value = max(value,0)
+        self.rgba1 = self.rgb1 + (value,)
+
+    #---spaces---
+    # Lingua franca is rgba
+    @property
+    def rgba(self):
+        if hasattr(self, '_franca'):
+            return self._franca
+    @rgba.setter
+    def rgba(self, color):
+        # Validate
+        color = self.validate(color, against=['rgb', 'rgba'], enforce=tuple)
+        if not color:
+            return
+        # Append alpha, if not present
+        if len(color) == 4:
+            self._franca = color
+        elif len(color) == 3:
+            self._franca = color + (1,)
+        else:
+            self.named = None
+        # Clear outdated values from cache
+        self._cache = {}
+
+    @property
+    def rgb(self):
+        if self.rgba:
+            return self.rgba[:-1]
+    @rgb.setter
+    def rgb(self, color):
+        self.rgba = color
+
+    @property
+    def rgba255(self):
+        if not self.rgba:
+            return None
+        if 'rgba255' not in self._cache:
+            self._cache['rgba255'] = tuple(int(255*(val+1)/2) for val in self.rgba)
+        return self._cache['rgba255']
+    @rgba255.setter
+    def rgba255(self, color):
+        # Validate
+        color = self.validate(color, against=['rgb255', 'rgba255'], enforce=(tuple, int))
+        if not color:
+            return
+        # Iterate through values and do conversion
+        self.rgba = tuple(2 * (val / 255 - 0.5) for val in color)
+        # Clear outdated values from cache
+        self._cache = {}
+
+    @property
+    def rgb255(self):
+        if self.rgba255:
+            return self.rgba255[:-1]
+    @rgb255.setter
+    def rgb255(self, color):
+        self.rgba255 = color
+
+    @property
+    def rgba1(self):
+        # Iterate through values and do conversion
+        if not self.rgba:
+            return
+        if 'rgba1' not in self._cache:
+            self._cache['rgba1'] = tuple((val + 1) / 2 for val in self.rgba)
+        return self._cache['rgba1']
+    @rgba1.setter
+    def rgba1(self, color):
+        # Validate
+        color = self.validate(color, against=['rgb1', 'rgba1'], enforce=tuple)
+        if not color:
+            return
+        # Iterate through values and do conversion
+        self.rgba = tuple(2 * (val - 0.5) for val in color)
+        # Clear outdated values from cache
+        self._cache = {}
+
+    @property
+    def rgb1(self):
+        if self.rgba1:
+            return self.rgba1[:-1]
+    @rgb1.setter
+    def rgb1(self, color):
+        self.rgba1 = color
+
+    @property
+    def hexa(self):
+        if not self.rgba255:
+            return
+        if 'hexa' not in self._cache:
+            # Map rgb255 values to corresponding letters in hex
+            hexmap = {10: 'a', 11: 'b', 12: 'c', 13: 'd', 14: 'e', 15: 'f'}
+            # Iterate and do conversion
+            flatList = ['#']
+            for val in self.rgba255:
+                dig1 = int(floor(val / 16))
+                flatList.append(
+                    str(dig1) if dig1 <= 9 else hexmap[dig1]
+                )
+                dig2 = int(val % 16)
+                flatList.append(
+                    str(dig2) if dig2 <= 9 else hexmap[dig2]
+                )
+            self._cache['hexa'] = "".join(flatList)
+        return self._cache['hexa']
+    @hexa.setter
+    def hexa(self, color):
+        # Validate
+        color = self.validate(color, against=['hex', 'hexa'], enforce=str)
+        if not color:
+            return
+        # Convert strings to list
+        colorList = [color[i - 2:i] for i in [3, 5, 7, 9] if color[i - 2:i]]
+        # Map hex letters to corresponding values in rgb255
+        hexmap = {'a':10, 'b':11, 'c':12, 'd':13, 'e':14, 'f':15}
+        # Create adjustment for different digits
+        adj = {0:16, 1:1}
+        flatList = []
+        for val in colorList:
+            # Iterate through individual values
+            flat = 0
+            for i, v in enumerate(val):
+                if re.match('\d', str(v)):
+                    flat += int(v)*adj[i]
+                elif re.match('[abcdef]', str(v).lower()):
+                    flat += hexmap[str(v).lower()]*adj[i]
+            flatList.append(flat)
+        self.rgba255 = flatList
+        # Clear outdated values from cache
+        self._cache = {}
+
+    @property
+    def hex(self):
+        if self.hexa:
+            return self.hexa[:-2]
+    @hex.setter
+    def hex(self, color):
+        self.hexa = color
+
+    @property
+    def named(self):
+        if 'named' not in self._cache:
+            # Round all values to 2 decimal places to find approximate matches
+            approxNames = {col: [round(val, 2) for val in colorNames[col]]
+                           for col in colorNames}
+            approxColor = [round(val, 2) for val in self.rgba]
+            # Get matches
+            possible = [nm for nm in approxNames if approxNames[nm] == approxColor]
+            # Return the first match
+            if possible:
+                self._cache['named'] = possible[0]
+            else:
+                self._cache['named'] = None
+        return self._cache['named']
+    @named.setter
+    def named(self, color):
+        # Validate
+        color = self.validate(color=str(color).lower(), against=['named'], enforce=str)
+        if not color:
+            return
+        # Retrieve named colour
+        self.rgba = colorNames[str(color).lower()]
+        # Clear outdated values from cache
+        self._cache = {}
+
+    @property
+    def hsva(self):
+        # Based on https://www.geeksforgeeks.org/program-change-rgb-color-model-hsv-color-model/
+        if 'hsva' not in self._cache:
+            red, green, blue, alpha = self.rgba1
+            cmax = max(red, green, blue)
+            cmin = min(red, green, blue)
+            delta = cmax - cmin
+            # Calculate hue
+            if cmax == 0 and cmin == 0:
+                return (0, 0, 0, alpha)
+            elif delta == 0:
+                return (0, 0, sum(self.rgb1)/3, alpha)
+
+            if cmax == red:
+                hue = (60 * ((green - blue) / delta) + 360) % 360
+            elif cmax == green:
+                hue = (60 * ((blue - red) / delta) + 120) % 360
+            elif cmax == blue:
+                hue = (60 * ((red - green) / delta) + 240) % 360
+            # Calculate saturation
+            if cmax == 0:
+                saturation = 0
+            else:
+                saturation = (delta / cmax)
+            # Calculate vibrancy
+            vibrancy = cmax
+            self._cache['hsva'] = (hue, saturation, vibrancy, alpha)
+        return self._cache['hsva']
+    @hsva.setter
+    def hsva(self, color):
+        # based on method in
+        # http://en.wikipedia.org/wiki/HSL_and_HSV#Converting_to_RGB
+
+        # Validate
+        color = self.validate(color, against=['hsva', 'hsv'], enforce=tuple)
+        if not color:
+            return
+        # Extract values
+        hue, saturation, vibrancy, *alpha = color
+        alpha255 = alpha[0] * 255 if alpha else None
+        # Convert hue
+        hue255 = Color.hue2rgb255(hue)
+        # Get value to move towards as saturation decreases
+        vibrancy255 = vibrancy*255
+        # Adjust by vibrancy and saturation
+        all255 = tuple(h+(vibrancy255-h)*(1-saturation) for h in hue255)
+        # Apply via rgba255
+        self.rgba255 = all255 + (alpha255,) if alpha255 is not None else all255 + (255,)
+        # Clear outdated values from cache
+        self._cache = {}
+    @property
+    def hsv(self):
+        if self.hsva:
+            return self.hsva[:-1]
+    @hsv.setter
+    def hsv(self, color):
+        self.hsva = color
+
+_rec = '(\-4\.5|\-4\.4\d*|\-4\.[0-4]\d*|\-[0-3]\.\d*|\-[0-3]|0|0\.\d*|1|1\.0)' # -4.5 to 1
+advancedSpaces = {
+    'rec709TF': re.compile(_lbr+_rec+',\s*'+_rec+',\s*'+_rec+_rbr), # rec709TF adjusted RGB from -4.5 to 1 + alpha from 0 to 1
+    'rec709TFa': re.compile(_lbr+_rec+',\s*'+_rec+',\s*'+_rec+',\s*'+_1+_rbr), # rec709TF adjusted RGB from -4.5 to 1 + alpha from 0 to 1
+    'srgbTF': re.compile(_lbr+'\-?'+_1+',\s*'+'\-?'+_1+',\s*'+'\-?'+_1+_rbr), # srgbTF from -1 to 1 + alpha from 0 to 1
+    'srgbTFa': re.compile(_lbr+'\-?'+_1+',\s*'+'\-?'+_1+',\s*'+'\-?'+_1+',\s*'+_1+_rbr), # srgbTF from -1 to 1 + alpha from 0 to 1
+    'lms': re.compile(_lbr + '\-?' + _1 + ',\s*' + '\-?' + _1 + ',\s*' + '\-?' + _1 + _rbr),  # LMS from -1 to 1
+    'lmsa': re.compile(_lbr + '\-?' + _1 + ',\s*' + '\-?' + _1 + ',\s*' + '\-?' + _1 + ',\s*' + '\-?' + _1 + _rbr),  # LMS + alpha from -1 to 1
+}
+
+
+class AdvancedColor(Color):
+    def __init__(self, color, space, conematrix=None):
+        Color.__init__(self, color, space)
+        # Set matrix for cone conversion
+        if conematrix is not None:
+            self.conematrix = conematrix
+        else:
+            # Set _conematrix specifically as undefined, rather than just setting to default
+            self._conematrix = None
+
+    @staticmethod
+    def getSpace(color, debug=False):
+        """Overrides Color.getSpace, drawing from a much more comprehensive library of colour spaces"""
+        if isinstance(color, Color):
+            return color._requestedSpace
+        # Check for advanced colours spaces
+        possible = [space for space in advancedSpaces
+                    if advancedSpaces[space].fullmatch(str(color))]
+        # Append basic colour spaces
+        basic = Color.getSpace(color, debug=True)
+        possible += basic
+        # Return full list if debug or multiple, else return first value
+        if debug or len(possible) > 1:
+            return possible
+        else:
+            return possible[0]
+
+    def validate(self, color, against=None, enforce=None):
+        # If not checking against anything, check against everything
+        if not against:
+            against = list(advancedSpaces)
+        # If looking for a string, convert from other forms it could be in
+        if enforce == str:
+            color = str(color).lower()
+        # If looking for a tuple, convert from other forms it could be in
+        if enforce == tuple:
+            if isinstance(color, str):
+                color = [float(n) for n in color.strip('[]()').split(',')]
+            if isinstance(color, list):
+                color = tuple(color)
+        # Get possible colour spaces
+        possible = AdvancedColor.getSpace(color)
+        if isinstance(possible, str):
+            possible = [possible]
+        # Return if any matches
+        for space in possible:
+            if space in against:
+                return color
+        # If no matches...
+        self.named = None
+        return None
+
+    @property
+    def rec709TFa(self):
+        """Apply the Rec. 709 transfer function (or gamma) to linear RGB values.
+
+            This transfer function is defined in the ITU-R BT.709 (2015) recommendation
+            document (http://www.itu.int/rec/R-REC-BT.709-6-201506-I/en) and is
+            commonly used with HDTV televisions.
+            """
+        if not self.rgb:
+            return
+        if 'rec709TFa' not in self._cache:
+            self._cache['rec709TFa'] = tuple(1.099 * c ** 0.45 - 0.099
+                         if c >= 0.018
+                         else 4.5 * c
+                         for c in self.rgb) + (self.rgba1[-1],)
+        return self._cache['rec709TFa']
+
+    @rec709TFa.setter
+    def rec709TFa(self, color):
+        # Validate
+        color = self.validate(color, against=['rec709TF', 'rec709TFa'], enforce=tuple)
+        if not color:
+            return
+        # Check for alpha
+        if len(color) == 4:
+            alpha = color[-1]
+            color = color[:-1]
+        elif len(color) == 3:
+            alpha = 1
+        # Do conversion
+        self.rgba = tuple(((c + 0.099)/1.099)**(1/0.45)
+                         if c >= 1.099 * 0.018 ** 0.45 - 0.099
+                         else c / 4.5
+                         for c in color) + (alpha,)
+        # Clear outdated values from cache
+        self._cache = {}
+
+    @property
+    def rec709TF(self):
+        if self.rec709TFa:
+            return self.rec709TFa[:-1]
+    @rec709TF.setter
+    def rec709TF(self, color):
+        self.rec709TFa = color
+
+    @property
+    def srgbTFa(self):
+        """Apply sRGB transfer function (or gamma) to linear RGB values."""
+        # applies the sRGB transfer function (linear RGB -> sRGB)
+        if not self.rgb:
+            return
+        if 'srgbTFa' not in self._cache:
+            self._cache['srgbTFa'] = tuple(c * 12.92
+                         if c <= 0.0031308
+                         else (1.0 + 0.055) * c ** (1.0 / 2.4) - 0.055
+                         for c in self.rgb) + (self.rgba1[-1],)
+        return self._cache['srgbTFa']
+    @srgbTFa.setter
+    def srgbTFa(self, color):
+        # Validate
+        color = self.validate(color, against=['srgbTF', 'srgbTFa'], enforce=tuple)
+        if not color:
+            return
+        # Check for alpha
+        if len(color) == 4:
+            alpha = color[-1]
+            color = color[:-1]
+        elif len(color) == 3:
+            alpha = 1
+        # do the inverse (sRGB -> linear RGB)
+        self.rgba = tuple(c / 12.92
+                     if c <= 0.04045
+                     else ((c + 0.055) / 1.055) ** 2.4
+                     for c in color) + (alpha,)
+        # Clear outdated values from cache
+        self._cache = {}
+
+    @property
+    def srgbTF(self):
+        if self.srgbTFa:
+            return self.srgbTFa[:-1]
+    @srgbTF.setter
+    def srgbTF(self, color):
+        self.srgbTFa = color
+
+    @property
+    def conematrix(self):
+        if self._conematrix is None:
+            # If _conematrix has been directly set to None, set to default
+            self.conematrix = None
+        return self._conematrix
+
+    @conematrix.setter
+    def conematrix(self, value):
+        # Default matrix
+        def default():
+            # Set default cone matrix and print warning
+            logging.warning('This monitor has not been color-calibrated. '
+                            'Using default LMS conversion matrix.')
+            return numpy.asarray([
+                # L        M        S
+                [4.97068857, -4.14354132, 0.17285275],  # R
+                [-0.90913894, 2.15671326, -0.24757432],  # G
+                [-0.03976551, -0.14253782, 1.18230333]])  # B
+
+        if not isinstance(value, numpy.ndarray):
+            self._conematrix = default()
+        elif not value.size == 9:
+            self._conematrix = default()
+        else:
+            self._conematrix = value
+
+    @property
+    def lmsa(self):
+        """Convert from RGB to cone space (LMS).
+
+        Requires a conversion matrix, which will be generated from generic
+        Sony Trinitron phosphors if not supplied (note that you will not get
+        an accurate representation of the color space unless you supply a
+        conversion matrix)
+        """
+        if 'lmsa' not in self._cache:
+            if self._conematrix is None:
+                self.conematrix = None
+            # its easier to use in the other orientation!
+            rgb_3xN = numpy.transpose(self.rgb)
+            rgb_to_cones = numpy.linalg.inv(self.conematrix)
+            lms = numpy.dot(rgb_to_cones, rgb_3xN)
+            self._cache['lmsa'] = tuple(numpy.transpose(lms))  # return in the shape we received it
+        return self._cache['lmsa']
+
+    @lmsa.setter
+    def lmsa(self, color):
+        """Convert from cone space (Long, Medium, Short) to RGB.
+
+        Requires a conversion matrix, which will be generated from generic
+        Sony Trinitron phosphors if not supplied (note that you will not get
+        an accurate representation of the color space unless you supply a
+        conversion matrix)
+        """
+        # Validate
+        color = self.validate(str(color).lower(), against=['lms', 'lmsa'], enforce=tuple)
+        if not color:
+            return
+        # Get alpha
+        if len(color) == 4:
+            alpha = color[-1]
+            color = color[:-1]
+        elif len(color) == 3:
+            alpha = 1
+        # its easier to use in the other orientation!
+        lms_3xN = numpy.transpose(color)
+        rgb = numpy.dot(self.conematrix, lms_3xN)
+        self.rgba = tuple(numpy.transpose(rgb)) + (alpha,)  # return in the shape we received it
+        # Clear outdated values from cache
+        self._cache = {}
+
+    @property
+    def lms(self):
+        if self.lmsa:
+            return self.lmsa[:-1]
+
+    @lms.setter
+    def lms(self, color):
+        self.lmsa = color
+
+"""----------Legacy-----------------"""
+# Old reference tables
+colors = colorNames
+colorsHex = {key: Color(key, 'named').hex for key in colors}
+colors255 = {key: Color(key, 'named').rgb255 for key in colors}
+
+# Old conversion functions
 def hex2rgb255(hexColor):
     """Convert a hex color string (e.g. "#05ff66") into an rgb triplet
     ranging from 0:255
@@ -17,7 +892,6 @@ def hex2rgb255(hexColor):
     elif len(hexColor.strip('#')) == 8:
         return col.rgba255
 
-
 def isValidColor(color):
     """check color validity (equivalent to existing checks in _setColor)
     """
@@ -26,447 +900,3 @@ def isValidColor(color):
         return True
     else:
         return False
-
-
-"""140 colors defined by most modern browsers (originally the standard colors
-of X11). Google for 140 web colors for further info"""
-colors = {
-    "aliceblue": (0.882352941176471, 0.945098039215686, 1),
-    "antiquewhite": (0.96078431372549, 0.843137254901961, 0.686274509803922),
-    "aqua": (-1, 1, 1),
-    "aquamarine": (-0.00392156862745097, 1, 0.662745098039216),
-    "azure": (0.882352941176471, 1, 1),
-    "beige": (0.92156862745098, 0.92156862745098, 0.725490196078431),
-    "bisque": (1, 0.788235294117647, 0.537254901960784),
-    "black": (-1, -1, -1),
-    "blanchedalmond": (1, 0.843137254901961, 0.607843137254902),
-    "blue": (-1, -1, 1),
-    "blueviolet": (0.0823529411764705, -0.662745098039216, 0.772549019607843),
-    "brown": (0.294117647058824, -0.670588235294118, -0.670588235294118),
-    "burlywood": (0.741176470588235, 0.443137254901961, 0.0588235294117647),
-    "cadetblue": (-0.254901960784314, 0.23921568627451, 0.254901960784314),
-    "chartreuse": (-0.00392156862745097, 1, -1),
-    "chocolate": (0.647058823529412, -0.176470588235294, -0.764705882352941),
-    "coral": (1, -0.00392156862745097, -0.372549019607843),
-    "cornflowerblue": (-0.215686274509804, 0.168627450980392, 0.858823529411765),
-    "cornsilk": (1, 0.945098039215686, 0.725490196078431),
-    "crimson": (0.725490196078431, -0.843137254901961, -0.529411764705882),
-    "cyan": (-1, 1, 1),
-    "darkblue": (-1, -1, 0.0901960784313725),
-    "darkcyan": (-1, 0.0901960784313725, 0.0901960784313725),
-    "darkgoldenrod": (0.443137254901961, 0.0509803921568628, -0.913725490196078),
-    "darkgray": (0.325490196078431, 0.325490196078431, 0.325490196078431),
-    "darkgreen": (-1, -0.215686274509804, -1),
-    "darkgrey": (0.325490196078431, 0.325490196078431, 0.325490196078431),
-    "darkkhaki": (0.482352941176471, 0.435294117647059, -0.16078431372549),
-    "darkmagenta": (0.0901960784313725, -1, 0.0901960784313725),
-    "darkolivegreen": (-0.333333333333333, -0.16078431372549, -0.631372549019608),
-    "darkorange": (1, 0.0980392156862746, -1),
-    "darkorchid": (0.2, -0.607843137254902, 0.6),
-    "darkred": (0.0901960784313725, -1, -1),
-    "darksalmon": (0.827450980392157, 0.176470588235294, -0.0431372549019607),
-    "darkseagreen": (0.12156862745098, 0.474509803921569, 0.12156862745098),
-    "darkslateblue": (-0.435294117647059, -0.52156862745098, 0.0901960784313725),
-    "darkslategray": (-0.631372549019608, -0.380392156862745, -0.380392156862745),
-    "darkslategrey": (-0.631372549019608, -0.380392156862745, -0.380392156862745),
-    "darkturquoise": (-1, 0.615686274509804, 0.63921568627451),
-    "darkviolet": (0.16078431372549, -1, 0.654901960784314),
-    "deeppink": (1, -0.843137254901961, 0.152941176470588),
-    "deepskyblue": (-1, 0.498039215686275, 1),
-    "dimgray": (-0.176470588235294, -0.176470588235294, -0.176470588235294),
-    "dimgrey": (-0.176470588235294, -0.176470588235294, -0.176470588235294),
-    "dodgerblue": (-0.764705882352941, 0.129411764705882, 1),
-    "firebrick": (0.396078431372549, -0.733333333333333, -0.733333333333333),
-    "floralwhite": (1, 0.96078431372549, 0.882352941176471),
-    "forestgreen": (-0.733333333333333, 0.0901960784313725, -0.733333333333333),
-    "fuchsia": (1, -1, 1),
-    "gainsboro": (0.725490196078431, 0.725490196078431, 0.725490196078431),
-    "ghostwhite": (0.945098039215686, 0.945098039215686, 1),
-    "gold": (1, 0.686274509803922, -1),
-    "goldenrod": (0.709803921568627, 0.294117647058824, -0.749019607843137),
-    "gray": (0.00392156862745097, 0.00392156862745097, 0.00392156862745097),
-    "grey": (0.00392156862745097, 0.00392156862745097, 0.00392156862745097),
-    "green": (-1, 0.00392156862745097, -1),
-    "greenyellow": (0.356862745098039, 1, -0.631372549019608),
-    "honeydew": (0.882352941176471, 1, 0.882352941176471),
-    "hotpink": (1, -0.176470588235294, 0.411764705882353),
-    "indianred": (0.607843137254902, -0.27843137254902, -0.27843137254902),
-    "indigo": (-0.411764705882353, -1, 0.0196078431372548),
-    "ivory": (1, 1, 0.882352941176471),
-    "khaki": (0.882352941176471, 0.803921568627451, 0.0980392156862746),
-    "lavender": (0.803921568627451, 0.803921568627451, 0.96078431372549),
-    "lavenderblush": (1, 0.882352941176471, 0.92156862745098),
-    "lawngreen": (-0.0274509803921569, 0.976470588235294, -1),
-    "lemonchiffon": (1, 0.96078431372549, 0.607843137254902),
-    "lightblue": (0.356862745098039, 0.694117647058824, 0.803921568627451),
-    "lightcoral": (0.882352941176471, 0.00392156862745097, 0.00392156862745097),
-    "lightcyan": (0.756862745098039, 1, 1),
-    "lightgoldenrodyellow": (0.96078431372549, 0.96078431372549, 0.647058823529412),
-    "lightgray": (0.654901960784314, 0.654901960784314, 0.654901960784314),
-    "lightgreen": (0.129411764705882, 0.866666666666667, 0.129411764705882),
-    "lightgrey": (0.654901960784314, 0.654901960784314, 0.654901960784314),
-    "lightpink": (1, 0.427450980392157, 0.513725490196078),
-    "lightsalmon": (1, 0.254901960784314, -0.0431372549019607),
-    "lightseagreen": (-0.749019607843137, 0.396078431372549, 0.333333333333333),
-    "lightskyblue": (0.0588235294117647, 0.615686274509804, 0.96078431372549),
-    "lightslategray": (-0.0666666666666667, 0.0666666666666667, 0.2),
-    "lightslategrey": (-0.0666666666666667, 0.0666666666666667, 0.2),
-    "lightsteelblue": (0.380392156862745, 0.537254901960784, 0.741176470588235),
-    "lightyellow": (1, 1, 0.756862745098039),
-    "lime": (-1, 1, -1),
-    "limegreen": (-0.607843137254902, 0.607843137254902, -0.607843137254902),
-    "linen": (0.96078431372549, 0.882352941176471, 0.803921568627451),
-    "magenta": (1, -1, 1),
-    "maroon": (0.00392156862745097, -1, -1),
-    "mediumaquamarine": (-0.2, 0.607843137254902, 0.333333333333333),
-    "mediumblue": (-1, -1, 0.607843137254902),
-    "mediumorchid": (0.458823529411765, -0.333333333333333, 0.654901960784314),
-    "mediumpurple": (0.152941176470588, -0.12156862745098, 0.717647058823529),
-    "mediumseagreen": (-0.529411764705882, 0.403921568627451, -0.113725490196078),
-    "mediumslateblue": (-0.0352941176470588, -0.184313725490196, 0.866666666666667),
-    "mediumspringgreen": (-1, 0.96078431372549, 0.207843137254902),
-    "mediumturquoise": (-0.435294117647059, 0.63921568627451, 0.6),
-    "mediumvioletred": (0.56078431372549, -0.835294117647059, 0.0431372549019609),
-    "midnightblue": (-0.803921568627451, -0.803921568627451, -0.12156862745098),
-    "mintcream": (0.92156862745098, 1, 0.96078431372549),
-    "mistyrose": (1, 0.788235294117647, 0.764705882352941),
-    "moccasin": (1, 0.788235294117647, 0.419607843137255),
-    "navajowhite": (1, 0.741176470588235, 0.356862745098039),
-    "navy": (-1, -1, 0.00392156862745097),
-    "oldlace": (0.984313725490196, 0.92156862745098, 0.803921568627451),
-    "olive": (0.00392156862745097, 0.00392156862745097, -1),
-    "olivedrab": (-0.16078431372549, 0.113725490196078, -0.725490196078431),
-    "orange": (1, 0.294117647058824, -1),
-    "orangered": (1, -0.458823529411765, -1),
-    "orchid": (0.709803921568627, -0.12156862745098, 0.67843137254902),
-    "palegoldenrod": (0.866666666666667, 0.819607843137255, 0.333333333333333),
-    "palegreen": (0.192156862745098, 0.968627450980392, 0.192156862745098),
-    "paleturquoise": (0.372549019607843, 0.866666666666667, 0.866666666666667),
-    "palevioletred": (0.717647058823529, -0.12156862745098, 0.152941176470588),
-    "papayawhip": (1, 0.874509803921569, 0.670588235294118),
-    "peachpuff": (1, 0.709803921568627, 0.450980392156863),
-    "peru": (0.607843137254902, 0.0431372549019609, -0.505882352941176),
-    "pink": (1, 0.505882352941176, 0.592156862745098),
-    "plum": (0.733333333333333, 0.254901960784314, 0.733333333333333),
-    "powderblue": (0.380392156862745, 0.756862745098039, 0.803921568627451),
-    "purple": (0.00392156862745097, -1, 0.00392156862745097),
-    "red": (1, -1, -1),
-    "rosybrown": (0.474509803921569, 0.12156862745098, 0.12156862745098),
-    "royalblue": (-0.490196078431373, -0.176470588235294, 0.764705882352941),
-    "saddlebrown": (0.0901960784313725, -0.458823529411765, -0.850980392156863),
-    "salmon": (0.96078431372549, 0.00392156862745097, -0.105882352941176),
-    "sandybrown": (0.913725490196079, 0.286274509803922, -0.247058823529412),
-    "seagreen": (-0.63921568627451, 0.0901960784313725, -0.317647058823529),
-    "seashell": (1, 0.92156862745098, 0.866666666666667),
-    "sienna": (0.254901960784314, -0.356862745098039, -0.647058823529412),
-    "silver": (0.505882352941176, 0.505882352941176, 0.505882352941176),
-    "skyblue": (0.0588235294117647, 0.615686274509804, 0.843137254901961),
-    "slateblue": (-0.168627450980392, -0.294117647058823, 0.607843137254902),
-    "slategray": (-0.12156862745098, 0.00392156862745097, 0.129411764705882),
-    "slategrey": (-0.12156862745098, 0.00392156862745097, 0.129411764705882),
-    "snow": (1, 0.96078431372549, 0.96078431372549),
-    "springgreen": (-1, 1, -0.00392156862745097),
-    "steelblue": (-0.450980392156863, 0.0196078431372548, 0.411764705882353),
-    "tan": (0.647058823529412, 0.411764705882353, 0.0980392156862746),
-    "teal": (-1, 0.00392156862745097, 0.00392156862745097),
-    "thistle": (0.694117647058824, 0.498039215686275, 0.694117647058824),
-    "tomato": (1, -0.223529411764706, -0.443137254901961),
-    "turquoise": (-0.498039215686275, 0.756862745098039, 0.631372549019608),
-    "violet": (0.866666666666667, 0.0196078431372548, 0.866666666666667),
-    "wheat": (0.92156862745098, 0.741176470588235, 0.403921568627451),
-    "white": (1, 1, 1),
-    "whitesmoke": (0.92156862745098, 0.92156862745098, 0.92156862745098),
-    "yellow": (1, 1, -1),
-    "yellowgreen": (0.207843137254902, 0.607843137254902, -0.607843137254902),
-}
-colorsHex = {
-    'aliceblue': '#F0F8FF',
-    'antiquewhite': '#FAEBD7',
-    'aqua': '#00FFFF',
-    'aquamarine': '#7FFFD4',
-    'azure': '#F0FFFF',
-    'beige': '#F5F5DC',
-    'bisque': '#FFE4C4',
-    'black': '#000000',
-    'blanchedalmond': '#FFEBCD',
-    'blue': '#0000FF',
-    'blueviolet': '#8A2BE2',
-    'brown': '#A52A2A',
-    'burlywood': '#DEB887',
-    'cadetblue': '#5F9EA0',
-    'chartreuse': '#7FFF00',
-    'chocolate': '#D2691E',
-    'coral': '#FF7F50',
-    'cornflowerblue': '#6495ED',
-    'cornsilk': '#FFF8DC',
-    'crimson': '#DC143C',
-    'cyan': '#00FFFF',
-    'darkblue': '#00008B',
-    'darkcyan': '#008B8B',
-    'darkgoldenrod': '#B8860B',
-    'darkgray': '#A9A9A9',
-    'darkgreen': '#006400',
-    'darkkhaki': '#BDB76B',
-    'darkmagenta': '#8B008B',
-    'darkolivegreen': '#556B2F',
-    'darkorange': '#FF8C00',
-    'darkorchid': '#9932CC',
-    'darkred': '#8B0000',
-    'darksalmon': '#E9967A',
-    'darkseagreen': '#8FBC8B',
-    'darkslateblue': '#483D8B',
-    'darkslategray': '#2F4F4F',
-    'darkturquoise': '#00CED1',
-    'darkviolet': '#9400D3',
-    'deeppink': '#FF1493',
-    'deepskyblue': '#00BFFF',
-    'dimgray': '#696969',
-    'dodgerblue': '#1E90FF',
-    'firebrick': '#B22222',
-    'floralwhite': '#FFFAF0',
-    'forestgreen': '#228B22',
-    'fuchsia': '#FF00FF',
-    'gainsboro': '#DCDCDC',
-    'ghostwhite': '#F8F8FF',
-    'gold': '#FFD700',
-    'goldenrod': '#DAA520',
-    'gray': '#808080',
-    'green': '#008000',
-    'greenyellow': '#ADFF2F',
-    'honeydew': '#F0FFF0',
-    'hotpink': '#FF69B4',
-    'indianred': '#CD5C5C',
-    'indigo': '#4B0082',
-    'ivory': '#FFFFF0',
-    'khaki': '#F0E68C',
-    'lavender': '#E6E6FA',
-    'lavenderblush': '#FFF0F5',
-    'lawngreen': '#7CFC00',
-    'lemonchiffon': '#FFFACD',
-    'lightblue': '#ADD8E6',
-    'lightcoral': '#F08080',
-    'lightcyan': '#E0FFFF',
-    'lightgoldenrodyellow': '#FAFAD2',
-    'lightgray': '#D3D3D3',
-    'lightgreen': '#90EE90',
-    'lightpink': '#FFB6C1',
-    'lightsalmon': '#FFA07A',
-    'lightseagreen': '#20B2AA',
-    'lightskyblue': '#87CEFA',
-    'lightslategray': '#778899',
-    'lightsteelblue': '#B0C4DE',
-    'lightyellow': '#FFFFE0',
-    'lime': '#00FF00',
-    'limegreen': '#32CD32',
-    'linen': '#FAF0E6',
-    'magenta': '#FF00FF',
-    'maroon': '#800000',
-    'mediumaquamarine': '#66CDAA',
-    'mediumblue': '#0000CD',
-    'mediumorchid': '#BA55D3',
-    'mediumpurple': '#9370DB',
-    'mediumseagreen': '#3CB371',
-    'mediumslateblue': '#7B68EE',
-    'mediumspringgreen': '#00FA9A',
-    'mediumturquoise': '#48D1CC',
-    'mediumvioletred': '#C71585',
-    'midnightblue': '#191970',
-    'mintcream': '#F5FFFA',
-    'mistyrose': '#FFE4E1',
-    'moccasin': '#FFE4B5',
-    'navajowhite': '#FFDEAD',
-    'navy': '#000080',
-    'oldlace': '#FDF5E6',
-    'olive': '#808000',
-    'olivedrab': '#6B8E23',
-    'orange': '#FFA500',
-    'orangered': '#FF4500',
-    'orchid': '#DA70D6',
-    'palegoldenrod': '#EEE8AA',
-    'palegreen': '#98FB98',
-    'paleturquoise': '#AFEEEE',
-    'palevioletred': '#DB7093',
-    'papayawhip': '#FFEFD5',
-    'peachpuff': '#FFDAB9',
-    'peru': '#CD853F',
-    'pink': '#FFC0CB',
-    'plum': '#DDA0DD',
-    'powderblue': '#B0E0E6',
-    'purple': '#800080',
-    'red': '#FF0000',
-    'rosybrown': '#BC8F8F',
-    'royalblue': '#4169E1',
-    'saddlebrown': '#8B4513',
-    'salmon': '#FA8072',
-    'sandybrown': '#F4A460',
-    'seagreen': '#2E8B57',
-    'seashell': '#FFF5EE',
-    'sienna': '#A0522D',
-    'silver': '#C0C0C0',
-    'skyblue': '#87CEEB',
-    'slateblue': '#6A5ACD',
-    'slategray': '#708090',
-    'snow': '#FFFAFA',
-    'springgreen': '#00FF7F',
-    'steelblue': '#4682B4',
-    'tan': '#D2B48C',
-    'teal': '#008080',
-    'thistle': '#D8BFD8',
-    'tomato': '#FF6347',
-    'turquoise': '#40E0D0',
-    'violet': '#EE82EE',
-    'wheat': '#F5DEB3',
-    'white': '#FFFFFF',
-    'whitesmoke': '#F5F5F5',
-    'yellow': '#FFFF00',
-    'yellowgreen': '#9ACD32'
-}
-colors255 = {
-    "aliceblue": (240, 248, 255),
-    "antiquewhite": (250, 235, 215),
-    "aqua": (0, 255, 255),
-    "aquamarine": (127, 255, 212),
-    "azure": (240, 255, 255),
-    "beige": (245, 245, 220),
-    "bisque": (255, 228, 196),
-    "black": (0, 0, 0),
-    "blanchedalmond": (255, 235, 205),
-    "blue": (0, 0, 255),
-    "blueviolet": (138, 43, 226),
-    "brown": (165, 42, 42),
-    "burlywood": (222, 184, 135),
-    "cadetblue": (95, 158, 160),
-    "chartreuse": (127, 255, 0),
-    "chocolate": (210, 105, 30),
-    "coral": (255, 127, 80),
-    "cornflowerblue": (100, 149, 237),
-    "cornsilk": (255, 248, 220),
-    "crimson": (220, 20, 60),
-    "cyan": (0, 255, 255),
-    "darkblue": (0, 0, 139),
-    "darkcyan": (0, 139, 139),
-    "darkgoldenrod": (184, 134, 11),
-    "darkgray": (169, 169, 169),
-    "darkgreen": (0, 100, 0),
-    "darkgrey": (169, 169, 169),
-    "darkkhaki": (189, 183, 107),
-    "darkmagenta": (139, 0, 139),
-    "darkolivegreen": (85, 107, 47),
-    "darkorange": (255, 140, 0),
-    "darkorchid": (153, 50, 204),
-    "darkred": (139, 0, 0),
-    "darksalmon": (233, 150, 122),
-    "darkseagreen": (143, 188, 143),
-    "darkslateblue": (72, 61, 139),
-    "darkslategray": (47, 79, 79),
-    "darkslategrey": (47, 79, 79),
-    "darkturquoise": (0, 206, 209),
-    "darkviolet": (148, 0, 211),
-    "deeppink": (255, 20, 147),
-    "deepskyblue": (0, 191, 255),
-    "dimgray": (105, 105, 105),
-    "dimgrey": (105, 105, 105),
-    "dodgerblue": (30, 144, 255),
-    "firebrick": (178, 34, 34),
-    "floralwhite": (255, 250, 240),
-    "forestgreen": (34, 139, 34),
-    "fuchsia": (255, 0, 255),
-    "gainsboro": (220, 220, 220),
-    "ghostwhite": (248, 248, 255),
-    "gold": (255, 215, 0),
-    "goldenrod": (218, 165, 32),
-    "gray": (128, 128, 128),
-    "grey": (128, 128, 128),
-    "green": (0, 128, 0),
-    "greenyellow": (173, 255, 47),
-    "honeydew": (240, 255, 240),
-    "hotpink": (255, 105, 180),
-    "indianred": (205, 92, 92),
-    "indigo": (75, 0, 130),
-    "ivory": (255, 255, 240),
-    "khaki": (240, 230, 140),
-    "lavender": (230, 230, 250),
-    "lavenderblush": (255, 240, 245),
-    "lawngreen": (124, 252, 0),
-    "lemonchiffon": (255, 250, 205),
-    "lightblue": (173, 216, 230),
-    "lightcoral": (240, 128, 128),
-    "lightcyan": (224, 255, 255),
-    "lightgoldenrodyellow": (250, 250, 210),
-    "lightgray": (211, 211, 211),
-    "lightgreen": (144, 238, 144),
-    "lightgrey": (211, 211, 211),
-    "lightpink": (255, 182, 193),
-    "lightsalmon": (255, 160, 122),
-    "lightseagreen": (32, 178, 170),
-    "lightskyblue": (135, 206, 250),
-    "lightslategray": (119, 136, 153),
-    "lightslategrey": (119, 136, 153),
-    "lightsteelblue": (176, 196, 222),
-    "lightyellow": (255, 255, 224),
-    "lime": (0, 255, 0),
-    "limegreen": (50, 205, 50),
-    "linen": (250, 240, 230),
-    "magenta": (255, 0, 255),
-    "maroon": (128, 0, 0),
-    "mediumaquamarine": (102, 205, 170),
-    "mediumblue": (0, 0, 205),
-    "mediumorchid": (186, 85, 211),
-    "mediumpurple": (147, 112, 219),
-    "mediumseagreen": (60, 179, 113),
-    "mediumslateblue": (123, 104, 238),
-    "mediumspringgreen": (0, 250, 154),
-    "mediumturquoise": (72, 209, 204),
-    "mediumvioletred": (199, 21, 133),
-    "midnightblue": (25, 25, 112),
-    "mintcream": (245, 255, 250),
-    "mistyrose": (255, 228, 225),
-    "moccasin": (255, 228, 181),
-    "navajowhite": (255, 222, 173),
-    "navy": (0, 0, 128),
-    "oldlace": (253, 245, 230),
-    "olive": (128, 128, 0),
-    "olivedrab": (107, 142, 35),
-    "orange": (255, 165, 0),
-    "orangered": (255, 69, 0),
-    "orchid": (218, 112, 214),
-    "palegoldenrod": (238, 232, 170),
-    "palegreen": (152, 251, 152),
-    "paleturquoise": (175, 238, 238),
-    "palevioletred": (219, 112, 147),
-    "papayawhip": (255, 239, 213),
-    "peachpuff": (255, 218, 185),
-    "peru": (205, 133, 63),
-    "pink": (255, 192, 203),
-    "plum": (221, 160, 221),
-    "powderblue": (176, 224, 230),
-    "purple": (128, 0, 128),
-    "red": (255, 0, 0),
-    "rosybrown": (188, 143, 143),
-    "royalblue": (65, 105, 225),
-    "saddlebrown": (139, 69, 19),
-    "salmon": (250, 128, 114),
-    "sandybrown": (244, 164, 96),
-    "seagreen": (46, 139, 87),
-    "seashell": (255, 245, 238),
-    "sienna": (160, 82, 45),
-    "silver": (192, 192, 192),
-    "skyblue": (135, 206, 235),
-    "slateblue": (106, 90, 205),
-    "slategray": (112, 128, 144),
-    "slategrey": (112, 128, 144),
-    "snow": (255, 250, 250),
-    "springgreen": (0, 255, 127),
-    "steelblue": (70, 130, 180),
-    "tan": (210, 180, 140),
-    "teal": (0, 128, 128),
-    "thistle": (216, 191, 216),
-    "tomato": (255, 99, 71),
-    "turquoise": (64, 224, 208),
-    "violet": (238, 130, 238),
-    "wheat": (245, 222, 179),
-    "white": (255, 255, 255),
-    "whitesmoke": (245, 245, 245),
-    "yellow": (255, 255, 0),
-    "yellowgreen": (154, 205, 50),
-}
