+++ psychopy/visual/basevisual.py
@@ -31,9 +31,6 @@ except ImportError:
 import copy
 import sys
 import os
-import re
-from math import floor, fsum
-
 from psychopy import logging
 
 # tools must only be imported *after* event or MovieStim breaks on win32
@@ -47,6 +44,8 @@ from psychopy.visual.helpers import (pointInPolygon, polygonsOverlap,
                                      setColor, findImageFile)
 from psychopy.tools.typetools import float_uint8
 from psychopy.tools.arraytools import makeRadialMatrix
+from psychopy.tools.colorspacetools import dkl2rgb, lms2rgb, hsv2rgb  # pylint: disable=W0611
+
 from . import globalVars
 
 import numpy
@@ -276,867 +275,6 @@ class LegacyVisualMixin(object):
         """
         self.__dict__['depth'] = value
 
-# Dict of examples of Psychopy Red at 12% opacity in different formats
-colorExamples = {
-    'invis': None,
-    'named': 'crimson',
-    'hex': '#F2545B',
-    'hexa': '#F2545B1E',
-    'rgb': (0.89, -0.35, -0.28),
-    'rgba': (0.89, -0.35, -0.28, -0.76),
-    'rgb1': (0.95, 0.32, 0.36),
-    'rgba1': (0.95, 0.32, 0.36, 0.12),
-    'rgb255': (242, 84, 91),
-    'rgba255': (242, 84, 91, 30),
-    'hsv': (357, 0.65, 0.95),
-    'hsva': (357, 0.65, 0.95, 0.12),
-}
-# Dict of named colours
-colorNames = {
-        "none": (0, 0, 0, -1),
-        "aliceblue": (0.882352941176471, 0.945098039215686, 1, 1.0),
-        "antiquewhite": (0.96078431372549, 0.843137254901961, 0.686274509803922, 1.0),
-        "aqua": (-1, 1, 1, 1.0),
-        "aquamarine": (-0.00392156862745097, 1, 0.662745098039216, 1.0),
-        "azure": (0.882352941176471, 1, 1, 1.0),
-        "beige": (0.92156862745098, 0.92156862745098, 0.725490196078431, 1.0),
-        "bisque": (1, 0.788235294117647, 0.537254901960784, 1.0),
-        "black": (-1, -1, -1, 1.0),
-        "blanchedalmond": (1, 0.843137254901961, 0.607843137254902, 1.0),
-        "blue": (-1, -1, 1, 1.0),
-        "blueviolet": (0.0823529411764705, -0.662745098039216, 0.772549019607843, 1.0),
-        "brown": (0.294117647058824, -0.670588235294118, -0.670588235294118, 1.0),
-        "burlywood": (0.741176470588235, 0.443137254901961, 0.0588235294117647, 1.0),
-        "cadetblue": (-0.254901960784314, 0.23921568627451, 0.254901960784314, 1.0),
-        "chartreuse": (-0.00392156862745097, 1, -1, 1.0),
-        "chocolate": (0.647058823529412, -0.176470588235294, -0.764705882352941, 1.0),
-        "coral": (1, -0.00392156862745097, -0.372549019607843, 1.0),
-        "cornflowerblue": (-0.215686274509804, 0.168627450980392, 0.858823529411765, 1.0),
-        "cornsilk": (1, 0.945098039215686, 0.725490196078431, 1.0),
-        "crimson": (0.725490196078431, -0.843137254901961, -0.529411764705882, 1.0),
-        "cyan": (-1, 1, 1, 1.0),
-        "darkblue": (-1, -1, 0.0901960784313725, 1.0),
-        "darkcyan": (-1, 0.0901960784313725, 0.0901960784313725, 1.0),
-        "darkgoldenrod": (0.443137254901961, 0.0509803921568628, -0.913725490196078, 1.0),
-        "darkgray": (0.325490196078431, 0.325490196078431, 0.325490196078431, 1.0),
-        "darkgreen": (-1, -0.215686274509804, -1, 1.0),
-        "darkgrey": (0.325490196078431, 0.325490196078431, 0.325490196078431, 1.0),
-        "darkkhaki": (0.482352941176471, 0.435294117647059, -0.16078431372549, 1.0),
-        "darkmagenta": (0.0901960784313725, -1, 0.0901960784313725, 1.0),
-        "darkolivegreen": (-0.333333333333333, -0.16078431372549, -0.631372549019608, 1.0),
-        "darkorange": (1, 0.0980392156862746, -1, 1.0),
-        "darkorchid": (0.2, -0.607843137254902, 0.6, 1.0),
-        "darkred": (0.0901960784313725, -1, -1, 1.0),
-        "darksalmon": (0.827450980392157, 0.176470588235294, -0.0431372549019607, 1.0),
-        "darkseagreen": (0.12156862745098, 0.474509803921569, 0.12156862745098, 1.0),
-        "darkslateblue": (-0.435294117647059, -0.52156862745098, 0.0901960784313725, 1.0),
-        "darkslategray": (-0.631372549019608, -0.380392156862745, -0.380392156862745, 1.0),
-        "darkslategrey": (-0.631372549019608, -0.380392156862745, -0.380392156862745, 1.0),
-        "darkturquoise": (-1, 0.615686274509804, 0.63921568627451, 1.0),
-        "darkviolet": (0.16078431372549, -1, 0.654901960784314, 1.0),
-        "deeppink": (1, -0.843137254901961, 0.152941176470588, 1.0),
-        "deepskyblue": (-1, 0.498039215686275, 1, 1.0),
-        "dimgray": (-0.176470588235294, -0.176470588235294, -0.176470588235294, 1.0),
-        "dimgrey": (-0.176470588235294, -0.176470588235294, -0.176470588235294, 1.0),
-        "dodgerblue": (-0.764705882352941, 0.129411764705882, 1, 1.0),
-        "firebrick": (0.396078431372549, -0.733333333333333, -0.733333333333333, 1.0),
-        "floralwhite": (1, 0.96078431372549, 0.882352941176471, 1.0),
-        "forestgreen": (-0.733333333333333, 0.0901960784313725, -0.733333333333333, 1.0),
-        "fuchsia": (1, -1, 1, 1.0),
-        "gainsboro": (0.725490196078431, 0.725490196078431, 0.725490196078431, 1.0),
-        "ghostwhite": (0.945098039215686, 0.945098039215686, 1, 1.0),
-        "gold": (1, 0.686274509803922, -1, 1.0),
-        "goldenrod": (0.709803921568627, 0.294117647058824, -0.749019607843137, 1.0),
-        "gray": (0.00392156862745097, 0.00392156862745097, 0.00392156862745097, 1.0),
-        "grey": (0.00392156862745097, 0.00392156862745097, 0.00392156862745097, 1.0),
-        "green": (-1, 0.00392156862745097, -1, 1.0),
-        "greenyellow": (0.356862745098039, 1, -0.631372549019608, 1.0),
-        "honeydew": (0.882352941176471, 1, 0.882352941176471, 1.0),
-        "hotpink": (1, -0.176470588235294, 0.411764705882353, 1.0),
-        "indianred": (0.607843137254902, -0.27843137254902, -0.27843137254902, 1.0),
-        "indigo": (-0.411764705882353, -1, 0.0196078431372548, 1.0),
-        "ivory": (1, 1, 0.882352941176471, 1.0),
-        "khaki": (0.882352941176471, 0.803921568627451, 0.0980392156862746, 1.0),
-        "lavender": (0.803921568627451, 0.803921568627451, 0.96078431372549, 1.0),
-        "lavenderblush": (1, 0.882352941176471, 0.92156862745098, 1.0),
-        "lawngreen": (-0.0274509803921569, 0.976470588235294, -1, 1.0),
-        "lemonchiffon": (1, 0.96078431372549, 0.607843137254902, 1.0),
-        "lightblue": (0.356862745098039, 0.694117647058824, 0.803921568627451, 1.0),
-        "lightcoral": (0.882352941176471, 0.00392156862745097, 0.00392156862745097, 1.0),
-        "lightcyan": (0.756862745098039, 1, 1, 1.0),
-        "lightgoldenrodyellow": (0.96078431372549, 0.96078431372549, 0.647058823529412, 1.0),
-        "lightgray": (0.654901960784314, 0.654901960784314, 0.654901960784314, 1.0),
-        "lightgreen": (0.129411764705882, 0.866666666666667, 0.129411764705882, 1.0),
-        "lightgrey": (0.654901960784314, 0.654901960784314, 0.654901960784314, 1.0),
-        "lightpink": (1, 0.427450980392157, 0.513725490196078, 1.0),
-        "lightsalmon": (1, 0.254901960784314, -0.0431372549019607, 1.0),
-        "lightseagreen": (-0.749019607843137, 0.396078431372549, 0.333333333333333, 1.0),
-        "lightskyblue": (0.0588235294117647, 0.615686274509804, 0.96078431372549, 1.0),
-        "lightslategray": (-0.0666666666666667, 0.0666666666666667, 0.2, 1.0),
-        "lightslategrey": (-0.0666666666666667, 0.0666666666666667, 0.2, 1.0),
-        "lightsteelblue": (0.380392156862745, 0.537254901960784, 0.741176470588235, 1.0),
-        "lightyellow": (1, 1, 0.756862745098039, 1.0),
-        "lime": (-1, 1, -1, 1.0),
-        "limegreen": (-0.607843137254902, 0.607843137254902, -0.607843137254902, 1.0),
-        "linen": (0.96078431372549, 0.882352941176471, 0.803921568627451, 1.0),
-        "magenta": (1, -1, 1, 1.0),
-        "maroon": (0.00392156862745097, -1, -1, 1.0),
-        "mediumaquamarine": (-0.2, 0.607843137254902, 0.333333333333333, 1.0),
-        "mediumblue": (-1, -1, 0.607843137254902, 1.0),
-        "mediumorchid": (0.458823529411765, -0.333333333333333, 0.654901960784314, 1.0),
-        "mediumpurple": (0.152941176470588, -0.12156862745098, 0.717647058823529, 1.0),
-        "mediumseagreen": (-0.529411764705882, 0.403921568627451, -0.113725490196078, 1.0),
-        "mediumslateblue": (-0.0352941176470588, -0.184313725490196, 0.866666666666667, 1.0),
-        "mediumspringgreen": (-1, 0.96078431372549, 0.207843137254902, 1.0),
-        "mediumturquoise": (-0.435294117647059, 0.63921568627451, 0.6, 1.0),
-        "mediumvioletred": (0.56078431372549, -0.835294117647059, 0.0431372549019609, 1.0),
-        "midnightblue": (-0.803921568627451, -0.803921568627451, -0.12156862745098, 1.0),
-        "mintcream": (0.92156862745098, 1, 0.96078431372549, 1.0),
-        "mistyrose": (1, 0.788235294117647, 0.764705882352941, 1.0),
-        "moccasin": (1, 0.788235294117647, 0.419607843137255, 1.0),
-        "navajowhite": (1, 0.741176470588235, 0.356862745098039, 1.0),
-        "navy": (-1, -1, 0.00392156862745097, 1.0),
-        "oldlace": (0.984313725490196, 0.92156862745098, 0.803921568627451, 1.0),
-        "olive": (0.00392156862745097, 0.00392156862745097, -1, 1.0),
-        "olivedrab": (-0.16078431372549, 0.113725490196078, -0.725490196078431, 1.0),
-        "orange": (1, 0.294117647058824, -1, 1.0),
-        "orangered": (1, -0.458823529411765, -1, 1.0),
-        "orchid": (0.709803921568627, -0.12156862745098, 0.67843137254902, 1.0),
-        "palegoldenrod": (0.866666666666667, 0.819607843137255, 0.333333333333333, 1.0),
-        "palegreen": (0.192156862745098, 0.968627450980392, 0.192156862745098, 1.0),
-        "paleturquoise": (0.372549019607843, 0.866666666666667, 0.866666666666667, 1.0),
-        "palevioletred": (0.717647058823529, -0.12156862745098, 0.152941176470588, 1.0),
-        "papayawhip": (1, 0.874509803921569, 0.670588235294118, 1.0),
-        "peachpuff": (1, 0.709803921568627, 0.450980392156863, 1.0),
-        "peru": (0.607843137254902, 0.0431372549019609, -0.505882352941176, 1.0),
-        "pink": (1, 0.505882352941176, 0.592156862745098, 1.0),
-        "plum": (0.733333333333333, 0.254901960784314, 0.733333333333333, 1.0),
-        "powderblue": (0.380392156862745, 0.756862745098039, 0.803921568627451, 1.0),
-        "purple": (0.00392156862745097, -1, 0.00392156862745097, 1.0),
-        "red": (1, -1, -1, 1.0),
-        "rosybrown": (0.474509803921569, 0.12156862745098, 0.12156862745098, 1.0),
-        "royalblue": (-0.490196078431373, -0.176470588235294, 0.764705882352941, 1.0),
-        "saddlebrown": (0.0901960784313725, -0.458823529411765, -0.850980392156863, 1.0),
-        "salmon": (0.96078431372549, 0.00392156862745097, -0.105882352941176, 1.0),
-        "sandybrown": (0.913725490196079, 0.286274509803922, -0.247058823529412, 1.0),
-        "seagreen": (-0.63921568627451, 0.0901960784313725, -0.317647058823529, 1.0),
-        "seashell": (1, 0.92156862745098, 0.866666666666667, 1.0),
-        "sienna": (0.254901960784314, -0.356862745098039, -0.647058823529412, 1.0),
-        "silver": (0.505882352941176, 0.505882352941176, 0.505882352941176, 1.0),
-        "skyblue": (0.0588235294117647, 0.615686274509804, 0.843137254901961, 1.0),
-        "slateblue": (-0.168627450980392, -0.294117647058823, 0.607843137254902, 1.0),
-        "slategray": (-0.12156862745098, 0.00392156862745097, 0.129411764705882, 1.0),
-        "slategrey": (-0.12156862745098, 0.00392156862745097, 0.129411764705882, 1.0),
-        "snow": (1, 0.96078431372549, 0.96078431372549, 1.0),
-        "springgreen": (-1, 1, -0.00392156862745097, 1.0),
-        "steelblue": (-0.450980392156863, 0.0196078431372548, 0.411764705882353, 1.0),
-        "tan": (0.647058823529412, 0.411764705882353, 0.0980392156862746, 1.0),
-        "teal": (-1, 0.00392156862745097, 0.00392156862745097, 1.0),
-        "thistle": (0.694117647058824, 0.498039215686275, 0.694117647058824, 1.0),
-        "tomato": (1, -0.223529411764706, -0.443137254901961, 1.0),
-        "turquoise": (-0.498039215686275, 0.756862745098039, 0.631372549019608, 1.0),
-        "violet": (0.866666666666667, 0.0196078431372548, 0.866666666666667, 1.0),
-        "wheat": (0.92156862745098, 0.741176470588235, 0.403921568627451, 1.0),
-        "white": (1, 1, 1, 1.0),
-        "whitesmoke": (0.92156862745098, 0.92156862745098, 0.92156862745098, 1.0),
-        "yellow": (1, 1, -1, 1.0),
-        "yellowgreen": (0.207843137254902, 0.607843137254902, -0.607843137254902, 1.0)
-    }
-# Shorthand for common regexpressions
-_255 = '(\d|\d\d|1\d\d|2[0-4]\d|25[0-5])'
-_360 = '(\d|\d\d|[12]\d\d|3[0-5]\d|360)'
-_100 = '(\d|\d\d|100)'
-_1 = '(0|1|1.0*|0\.\d*)'
-_lbr = '[\[\(]\s*'
-_rbr = '\s*[\]\)]'
-# Dict of regexpressions for different formats
-colorSpaces = {
-    'named': re.compile("|".join(list(colorNames))), # A named colour space
-    'hex': re.compile('#[\dabcdefABCDEF]{6}'), # Hex
-    'hexa': re.compile('#[\dabcdefABCDEF]{8}'), # Hex + alpha
-    'rgb': re.compile(_lbr+'\-?'+_1+',\s*'+'\-?'+_1+',\s*'+'\-?'+_1+_rbr), # RGB from -1 to 1
-    'rgba': re.compile(_lbr+'\-?'+_1+',\s*'+'\-?'+_1+',\s*'+'\-?'+_1+',\s*'+'\-?'+_1+_rbr),  # RGB + alpha from -1 to 1
-    'rgb1': re.compile(_lbr+_1+',\s*'+_1+',\s*'+_1+_rbr),  # RGB from 0 to 1
-    'rgba1': re.compile(_lbr+_1+',\s*'+_1+',\s*'+_1+',\s*'+_1+_rbr),  # RGB + alpha from 0 to 1
-    'rgb255': re.compile(_lbr+_255+',\s*'+_255+',\s*'+_255+_rbr), # RGB from 0 to 255
-    'rgba255': re.compile(_lbr+_255+',\s*'+_255+',\s*'+_255+',\s*'+_255+_rbr), # RGB + alpha from 0 to 255
-    'hsv': re.compile(_lbr+_360+'\°?'+',\s*'+_1+',\s*'+_1+_rbr), # HSV with hue from 0 to 360 and saturation/vibrancy from 0 to 1
-    'hsva': re.compile(_lbr+_360+'\°?'+',\s*'+_1+',\s*'+_1+',\s*'+_1+_rbr), # HSV with hue from 0 to 360 and saturation/vibrancy from 0 to 1 + alpha from 0 to 1
-}
-
-
-class Color(object):
-    """A class to store colour details, knows what colour space it's in and can supply colours in any space"""
-
-    def __init__(self, color=None, space=None):
-        self._cache = {}
-        self.set(color=color, space=space)
-
-    def set(self, color=None, space=None):
-        """Set the colour of this object - essentially the same as what happens on creation, but without
-        having to initialise a new object"""
-        if isinstance(color, numpy.ndarray):
-            color = tuple(float(c) for c in color)
-        if space in ['rgb255', 'rgba255']:
-            color = tuple(int(c) for c in color)
-        # If input is a Color object, duplicate all settings
-        if isinstance(color, Color):
-            self._requested = color._requested
-            self._requestedSpace = color._requestedSpace
-            self.rgba = color.rgba
-            return
-        # Store requested colour and space (or defaults, if none given)
-        self._requested = color if color is not None else None
-        self._requestedSpace = space \
-            if space and space in self.getSpace(self._requested, debug=True) \
-            else self.getSpace(self._requested)
-        if isinstance(self._requestedSpace, (list, type(None))):
-            logging.error("Color space could not be determined by values supplied, please specify a color space.")
-            return
-
-        # Convert to lingua franca
-        if self._requestedSpace:
-            setattr(self, self._requestedSpace, self._requested)
-        else:
-            self.named = None
-
-    def __repr__(self):
-        """If colour is printed, it will display its class and value"""
-        if self.rgba:
-            if self.named:
-                return "<" + self.__class__.__module__ + "." + self.__class__.__name__ + ": " + self.named + ">"
-            else:
-                return "<" + self.__class__.__module__ + "." + self.__class__.__name__ + ": " + str(tuple(round(c,2) for c in self.rgba)) + ">"
-        else:
-            return "<" + self.__class__.__module__ + "." + self.__class__.__name__ + ": " + "Invalid" + ">"
-
-    # ---rich comparisons---
-    def __eq__(self, target):
-        """== will compare RGBA values, rounded to 2dp"""
-        if isinstance(target, Color):
-            return (round(c,2) for c in self.rgba) == (round(c,2) for c in target.rgba)
-        else:
-            return False
-    def __ne__(self, target):
-        """!= will return the opposite of =="""
-        return not self == target
-    def __lt__(self, target):
-        """< will compare brightness"""
-        if isinstance(target, Color):
-            return self.brightness < target.brightness
-        else:
-            return False
-    def __le__(self, target):
-        """<= will compare brightness"""
-        if isinstance(target, Color):
-            return self.brightness <= target.brightness
-        else:
-            return False
-    def __gt__(self, target):
-        """> will compare brightness"""
-        if isinstance(target, Color):
-            return self.brightness > target.brightness
-        else:
-            return False
-    def __ge__(self, target):
-        """>= will compare brightness"""
-        if isinstance(target, Color):
-            return self.brightness >= target.brightness
-        else:
-            return False
-
-    #--operators---
-    def __add__(self, other):
-        if not isinstance(other, Color):
-            if Color.getSpace(other):
-                # Convert to a color if valid
-                other = Color(other)
-            elif AdvancedColor.getSpace(other):
-                # Convert to an advanced color if valid
-                other = AdvancedColor(other)
-            elif isinstance(other, int) or isinstance(other, float):
-                out = self.copy()
-                out.brightness += other
-                return out
-            else:
-                raise ValueError ("unsupported operand type(s) for +: '"
-                                  + self.__class__.__name__ +"' and '"
-                                  + other.__class__.__name__ + "'")
-        if isinstance(other, Color):
-            # If both are colours, average the two and sum their alphas
-            alpha = min(self.alpha + other.alpha, 1)
-            rgb = [None, None, None]
-            selfWeight = self.alpha/(self.alpha+other.alpha)
-            otherWeight = other.alpha/(self.alpha+other.alpha)
-            for c in range(3):
-                rgb[c] = self.rgb1[c]*selfWeight + other.rgb1[c]*otherWeight
-            return Color(rgb+[alpha], 'rgba1')
-
-
-    def copy(self):
-        """Return a duplicate of this colour"""
-        return self.__class__(self._requested, self._requestedSpace)
-
-    @staticmethod
-    def getSpace(color, debug=False):
-        """Find what colour space a colour is from"""
-        if isinstance(color, Color):
-            return color._requestedSpace
-        possible = [space for space in colorSpaces
-                    if colorSpaces[space].fullmatch(str(color).lower())]
-        # Return full list if debug or multiple, else return first value
-        if debug or not len(possible) == 1:
-            return possible
-        else:
-            return possible[0]
-
-    @staticmethod
-    def hue2rgb255(hue):
-        # Work out what segment of the colour wheel we're in
-        seg = floor(hue / 60)
-        seg = seg if seg % 6 else 0
-        # Define values for when a value is decreasing / increasing in a segment
-        _up = (hue % 60) * (255 / 60)
-        _down = 255 - _up
-        _mov = _down if seg%2 else _up # Even segments are down, odd are up
-        # Calculate rgb according to segment
-        if seg == 0:
-            return (255, _mov, 0,)
-        if seg == 1:
-            return (_mov, 255, 0,)
-        if seg == 2:
-            return (0, 255, _mov,)
-        if seg == 3:
-            return (0, _mov, 255,)
-        if seg == 4:
-            return (_mov, 0, 255,)
-        if seg == 5:
-            return (255, 0, _mov,)
-
-    def validate(self, color, against=None, enforce=None):
-        # If not checking against anything, check against everything
-        if not against:
-            against = list(colorSpaces)
-        # If looking for a string, convert from other forms it could be in
-        if enforce == str:
-            color = str(color).lower()
-        # If looking for a tuple, convert from other forms it could be in
-        if enforce == tuple or enforce == (tuple, int):
-            if isinstance(color, str):
-                color = [float(n) for n in color.strip('[]()').split(',')]
-            if isinstance(color, list):
-                color = tuple(color)
-        # If enforcing multiple
-        if enforce == (tuple, int):
-            color = tuple(int(round(c)) for c in color)
-        # Get possible colour spaces
-        possible = Color.getSpace(color)
-        if isinstance(possible, str):
-            possible = [possible]
-        # Return if any matches
-        for space in possible:
-            if space in against:
-                return color
-        # If no matches...
-        self.named = None
-        return None
-
-    # ---adjusters---
-    @property
-    def contrast(self):
-        '''Distance from mid grey in rgb (-1 to 1)'''
-        return abs(sum(self.rgb)/3)
-    @contrast.setter
-    def contrast(self, value):
-        if self.contrast:
-            norm = tuple(c/self.contrast for c in self.rgb)
-        else:
-            norm = self.rgb
-        self.rgba = tuple(max(min(c*value,1),-1) for c in norm) + (self.rgba[-1],)
-
-    @property
-    def saturation(self):
-        '''Saturation in hsv (0 to 1)'''
-        return self.hsv[1]
-    @saturation.setter
-    def saturation(self, value):
-        h,s,v,a = self.hsva
-        s += value
-        s = min(s, 1)
-        s = max(s, 0)
-        self.hsva = (h, s, v, a)
-
-    @property
-    def brightness(self):
-        return sum(self.rgb1)/3
-    @brightness.setter
-    def brightness(self, value):
-        adj = value-self.brightness
-        self.rgba1 = tuple(
-            max(min(c+adj, 1),0)
-            for c in self.rgb1
-        ) + (self.alpha,)
-
-    @property
-    def alpha(self):
-        return self.rgba1[-1]
-    @alpha.setter
-    def alpha(self, value):
-        value = min(value,1)
-        value = max(value,0)
-        self.rgba1 = self.rgb1 + (value,)
-
-    #---spaces---
-    # Lingua franca is rgba
-    @property
-    def rgba(self):
-        if hasattr(self, '_franca'):
-            return self._franca
-    @rgba.setter
-    def rgba(self, color):
-        # Validate
-        color = self.validate(color, against=['rgb', 'rgba'], enforce=tuple)
-        if not color:
-            return
-        # Append alpha, if not present
-        if len(color) == 4:
-            self._franca = color
-        elif len(color) == 3:
-            self._franca = color + (1,)
-        else:
-            self.named = None
-        # Clear outdated values from cache
-        self._cache = {}
-
-    @property
-    def rgb(self):
-        if self.rgba:
-            return self.rgba[:-1]
-    @rgb.setter
-    def rgb(self, color):
-        self.rgba = color
-
-    @property
-    def rgba255(self):
-        if not self.rgba:
-            return None
-        if 'rgba255' not in self._cache:
-            self._cache['rgba255'] = tuple(int(255*(val+1)/2) for val in self.rgba)
-        return self._cache['rgba255']
-    @rgba255.setter
-    def rgba255(self, color):
-        # Validate
-        color = self.validate(color, against=['rgb255', 'rgba255'], enforce=(tuple, int))
-        if not color:
-            return
-        # Iterate through values and do conversion
-        self.rgba = tuple(2 * (val / 255 - 0.5) for val in color)
-        # Clear outdated values from cache
-        self._cache = {}
-
-    @property
-    def rgb255(self):
-        if self.rgba255:
-            return self.rgba255[:-1]
-    @rgb255.setter
-    def rgb255(self, color):
-        self.rgba255 = color
-
-    @property
-    def rgba1(self):
-        # Iterate through values and do conversion
-        if not self.rgba:
-            return
-        if 'rgba1' not in self._cache:
-            self._cache['rgba1'] = tuple((val + 1) / 2 for val in self.rgba)
-        return self._cache['rgba1']
-    @rgba1.setter
-    def rgba1(self, color):
-        # Validate
-        color = self.validate(color, against=['rgb1', 'rgba1'], enforce=tuple)
-        if not color:
-            return
-        # Iterate through values and do conversion
-        self.rgba = tuple(2 * (val - 0.5) for val in color)
-        # Clear outdated values from cache
-        self._cache = {}
-
-    @property
-    def rgb1(self):
-        if self.rgba1:
-            return self.rgba1[:-1]
-    @rgb1.setter
-    def rgb1(self, color):
-        self.rgba1 = color
-
-    @property
-    def hexa(self):
-        if not self.rgba255:
-            return
-        if 'hexa' not in self._cache:
-            # Map rgb255 values to corresponding letters in hex
-            hexmap = {10: 'a', 11: 'b', 12: 'c', 13: 'd', 14: 'e', 15: 'f'}
-            # Iterate and do conversion
-            flatList = ['#']
-            for val in self.rgba255:
-                dig1 = int(floor(val / 16))
-                flatList.append(
-                    str(dig1) if dig1 <= 9 else hexmap[dig1]
-                )
-                dig2 = int(val % 16)
-                flatList.append(
-                    str(dig2) if dig2 <= 9 else hexmap[dig2]
-                )
-            self._cache['hexa'] = "".join(flatList)
-        return self._cache['hexa']
-    @hexa.setter
-    def hexa(self, color):
-        # Validate
-        color = self.validate(color, against=['hex', 'hexa'], enforce=str)
-        if not color:
-            return
-        # Convert strings to list
-        colorList = [color[i - 2:i] for i in [3, 5, 7, 9] if color[i - 2:i]]
-        # Map hex letters to corresponding values in rgb255
-        hexmap = {'a':10, 'b':11, 'c':12, 'd':13, 'e':14, 'f':15}
-        # Create adjustment for different digits
-        adj = {0:16, 1:1}
-        flatList = []
-        for val in colorList:
-            # Iterate through individual values
-            flat = 0
-            for i, v in enumerate(val):
-                if re.match('\d', str(v)):
-                    flat += int(v)*adj[i]
-                elif re.match('[abcdef]', str(v).lower()):
-                    flat += hexmap[str(v).lower()]*adj[i]
-            flatList.append(flat)
-        self.rgba255 = flatList
-        # Clear outdated values from cache
-        self._cache = {}
-
-    @property
-    def hex(self):
-        if self.hexa:
-            return self.hexa[:-2]
-    @hex.setter
-    def hex(self, color):
-        self.hexa = color
-
-    @property
-    def named(self):
-        if 'named' not in self._cache:
-            # Round all values to 2 decimal places to find approximate matches
-            approxNames = {col: [round(val, 2) for val in colorNames[col]]
-                           for col in colorNames}
-            approxColor = [round(val, 2) for val in self.rgba]
-            # Get matches
-            possible = [nm for nm in approxNames if approxNames[nm] == approxColor]
-            # Return the first match
-            if possible:
-                self._cache['named'] = possible[0]
-            else:
-                self._cache['named'] = None
-        return self._cache['named']
-    @named.setter
-    def named(self, color):
-        # Validate
-        color = self.validate(color=str(color).lower(), against=['named'], enforce=str)
-        if not color:
-            return
-        # Retrieve named colour
-        self.rgba = colorNames[str(color).lower()]
-        # Clear outdated values from cache
-        self._cache = {}
-
-    @property
-    def hsva(self):
-        # Based on https://www.geeksforgeeks.org/program-change-rgb-color-model-hsv-color-model/
-        if 'hsva' not in self._cache:
-            red, green, blue, alpha = self.rgba1
-            cmax = max(red, green, blue)
-            cmin = min(red, green, blue)
-            delta = cmax - cmin
-            # Calculate hue
-            if cmax == 0 and cmin == 0:
-                return (0, 0, 0, alpha)
-            elif delta == 0:
-                return (0, 0, sum(self.rgb1)/3, alpha)
-
-            if cmax == red:
-                hue = (60 * ((green - blue) / delta) + 360) % 360
-            elif cmax == green:
-                hue = (60 * ((blue - red) / delta) + 120) % 360
-            elif cmax == blue:
-                hue = (60 * ((red - green) / delta) + 240) % 360
-            # Calculate saturation
-            if cmax == 0:
-                saturation = 0
-            else:
-                saturation = (delta / cmax)
-            # Calculate vibrancy
-            vibrancy = cmax
-            self._cache['hsva'] = (hue, saturation, vibrancy, alpha)
-        return self._cache['hsva']
-    @hsva.setter
-    def hsva(self, color):
-        # based on method in
-        # http://en.wikipedia.org/wiki/HSL_and_HSV#Converting_to_RGB
-
-        # Validate
-        color = self.validate(color, against=['hsva', 'hsv'], enforce=tuple)
-        if not color:
-            return
-        # Extract values
-        hue, saturation, vibrancy, *alpha = color
-        alpha255 = alpha[0] * 255 if alpha else None
-        # Convert hue
-        hue255 = Color.hue2rgb255(hue)
-        # Get value to move towards as saturation decreases
-        vibrancy255 = vibrancy*255
-        # Adjust by vibrancy and saturation
-        all255 = tuple(h+(vibrancy255-h)*(1-saturation) for h in hue255)
-        # Apply via rgba255
-        self.rgba255 = all255 + (alpha255,) if alpha255 is not None else all255 + (255,)
-        # Clear outdated values from cache
-        self._cache = {}
-    @property
-    def hsv(self):
-        if self.hsva:
-            return self.hsva[:-1]
-    @hsv.setter
-    def hsv(self, color):
-        self.hsva = color
-
-_rec = '(\-4\.5|\-4\.4\d*|\-4\.[0-4]\d*|\-[0-3]\.\d*|\-[0-3]|0|0\.\d*|1|1\.0)' # -4.5 to 1
-advancedSpaces = {
-    'rec709TF': re.compile(_lbr+_rec+',\s*'+_rec+',\s*'+_rec+_rbr), # rec709TF adjusted RGB from -4.5 to 1 + alpha from 0 to 1
-    'rec709TFa': re.compile(_lbr+_rec+',\s*'+_rec+',\s*'+_rec+',\s*'+_1+_rbr), # rec709TF adjusted RGB from -4.5 to 1 + alpha from 0 to 1
-    'srgbTF': re.compile(_lbr+'\-?'+_1+',\s*'+'\-?'+_1+',\s*'+'\-?'+_1+_rbr), # srgbTF from -1 to 1 + alpha from 0 to 1
-    'srgbTFa': re.compile(_lbr+'\-?'+_1+',\s*'+'\-?'+_1+',\s*'+'\-?'+_1+',\s*'+_1+_rbr), # srgbTF from -1 to 1 + alpha from 0 to 1
-    'lms': re.compile(_lbr + '\-?' + _1 + ',\s*' + '\-?' + _1 + ',\s*' + '\-?' + _1 + _rbr),  # LMS from -1 to 1
-    'lmsa': re.compile(_lbr + '\-?' + _1 + ',\s*' + '\-?' + _1 + ',\s*' + '\-?' + _1 + ',\s*' + '\-?' + _1 + _rbr),  # LMS + alpha from -1 to 1
-}
-
-
-class AdvancedColor(Color):
-    def __init__(self, color, space, conematrix=None):
-        Color.__init__(self, color, space)
-        # Set matrix for cone conversion
-        if conematrix is not None:
-            self.conematrix = conematrix
-        else:
-            # Set _conematrix specifically as undefined, rather than just setting to default
-            self._conematrix = None
-
-    @staticmethod
-    def getSpace(color, debug=False):
-        """Overrides Color.getSpace, drawing from a much more comprehensive library of colour spaces"""
-        if isinstance(color, Color):
-            return color._requestedSpace
-        # Check for advanced colours spaces
-        possible = [space for space in advancedSpaces
-                    if advancedSpaces[space].fullmatch(str(color))]
-        # Append basic colour spaces
-        basic = Color.getSpace(color, debug=True)
-        possible += basic
-        # Return full list if debug or multiple, else return first value
-        if debug or len(possible) > 1:
-            return possible
-        else:
-            return possible[0]
-
-    def validate(self, color, against=None, enforce=None):
-        # If not checking against anything, check against everything
-        if not against:
-            against = list(advancedSpaces)
-        # If looking for a string, convert from other forms it could be in
-        if enforce == str:
-            color = str(color).lower()
-        # If looking for a tuple, convert from other forms it could be in
-        if enforce == tuple:
-            if isinstance(color, str):
-                color = [float(n) for n in color.strip('[]()').split(',')]
-            if isinstance(color, list):
-                color = tuple(color)
-        # Get possible colour spaces
-        possible = AdvancedColor.getSpace(color)
-        if isinstance(possible, str):
-            possible = [possible]
-        # Return if any matches
-        for space in possible:
-            if space in against:
-                return color
-        # If no matches...
-        self.named = None
-        return None
-
-    @property
-    def rec709TFa(self):
-        """Apply the Rec. 709 transfer function (or gamma) to linear RGB values.
-
-            This transfer function is defined in the ITU-R BT.709 (2015) recommendation
-            document (http://www.itu.int/rec/R-REC-BT.709-6-201506-I/en) and is
-            commonly used with HDTV televisions.
-            """
-        if not self.rgb:
-            return
-        if 'rec709TFa' not in self._cache:
-            self._cache['rec709TFa'] = tuple(1.099 * c ** 0.45 - 0.099
-                         if c >= 0.018
-                         else 4.5 * c
-                         for c in self.rgb) + (self.rgba1[-1],)
-        return self._cache['rec709TFa']
-
-    @rec709TFa.setter
-    def rec709TFa(self, color):
-        # Validate
-        color = self.validate(color, against=['rec709TF', 'rec709TFa'], enforce=tuple)
-        if not color:
-            return
-        # Check for alpha
-        if len(color) == 4:
-            alpha = color[-1]
-            color = color[:-1]
-        elif len(color) == 3:
-            alpha = 1
-        # Do conversion
-        self.rgba = tuple(((c + 0.099)/1.099)**(1/0.45)
-                         if c >= 1.099 * 0.018 ** 0.45 - 0.099
-                         else c / 4.5
-                         for c in color) + (alpha,)
-        # Clear outdated values from cache
-        self._cache = {}
-
-    @property
-    def rec709TF(self):
-        if self.rec709TFa:
-            return self.rec709TFa[:-1]
-    @rec709TF.setter
-    def rec709TF(self, color):
-        self.rec709TFa = color
-
-    @property
-    def srgbTFa(self):
-        """Apply sRGB transfer function (or gamma) to linear RGB values."""
-        # applies the sRGB transfer function (linear RGB -> sRGB)
-        if not self.rgb:
-            return
-        if 'srgbTFa' not in self._cache:
-            self._cache['srgbTFa'] = tuple(c * 12.92
-                         if c <= 0.0031308
-                         else (1.0 + 0.055) * c ** (1.0 / 2.4) - 0.055
-                         for c in self.rgb) + (self.rgba1[-1],)
-        return self._cache['srgbTFa']
-    @srgbTFa.setter
-    def srgbTFa(self, color):
-        # Validate
-        color = self.validate(color, against=['srgbTF', 'srgbTFa'], enforce=tuple)
-        if not color:
-            return
-        # Check for alpha
-        if len(color) == 4:
-            alpha = color[-1]
-            color = color[:-1]
-        elif len(color) == 3:
-            alpha = 1
-        # do the inverse (sRGB -> linear RGB)
-        self.rgba = tuple(c / 12.92
-                     if c <= 0.04045
-                     else ((c + 0.055) / 1.055) ** 2.4
-                     for c in color) + (alpha,)
-        # Clear outdated values from cache
-        self._cache = {}
-
-    @property
-    def srgbTF(self):
-        if self.srgbTFa:
-            return self.srgbTFa[:-1]
-    @srgbTF.setter
-    def srgbTF(self, color):
-        self.srgbTFa = color
-
-    @property
-    def conematrix(self):
-        if self._conematrix is None:
-            # If _conematrix has been directly set to None, set to default
-            self.conematrix = None
-        return self._conematrix
-
-    @conematrix.setter
-    def conematrix(self, value):
-        # Default matrix
-        def default():
-            # Set default cone matrix and print warning
-            logging.warning('This monitor has not been color-calibrated. '
-                            'Using default LMS conversion matrix.')
-            return numpy.asarray([
-                # L        M        S
-                [4.97068857, -4.14354132, 0.17285275],  # R
-                [-0.90913894, 2.15671326, -0.24757432],  # G
-                [-0.03976551, -0.14253782, 1.18230333]])  # B
-
-        if not isinstance(value, numpy.ndarray):
-            self._conematrix = default()
-        elif not value.size == 9:
-            self._conematrix = default()
-        else:
-            self._conematrix = value
-
-    @property
-    def lmsa(self):
-        """Convert from RGB to cone space (LMS).
-
-        Requires a conversion matrix, which will be generated from generic
-        Sony Trinitron phosphors if not supplied (note that you will not get
-        an accurate representation of the color space unless you supply a
-        conversion matrix)
-        """
-        if 'lmsa' not in self._cache:
-            if self._conematrix is None:
-                self.conematrix = None
-            # its easier to use in the other orientation!
-            rgb_3xN = numpy.transpose(self.rgb)
-            rgb_to_cones = numpy.linalg.inv(self.conematrix)
-            lms = numpy.dot(rgb_to_cones, rgb_3xN)
-            self._cache['lmsa'] = tuple(numpy.transpose(lms))  # return in the shape we received it
-        return self._cache['lmsa']
-
-    @lmsa.setter
-    def lmsa(self, color):
-        """Convert from cone space (Long, Medium, Short) to RGB.
-
-        Requires a conversion matrix, which will be generated from generic
-        Sony Trinitron phosphors if not supplied (note that you will not get
-        an accurate representation of the color space unless you supply a
-        conversion matrix)
-        """
-        # Validate
-        color = self.validate(str(color).lower(), against=['lms', 'lmsa'], enforce=tuple)
-        if not color:
-            return
-        # Get alpha
-        if len(color) == 4:
-            alpha = color[-1]
-            color = color[:-1]
-        elif len(color) == 3:
-            alpha = 1
-        # its easier to use in the other orientation!
-        lms_3xN = numpy.transpose(color)
-        rgb = numpy.dot(self.conematrix, lms_3xN)
-        self.rgba = tuple(numpy.transpose(rgb)) + (alpha,)  # return in the shape we received it
-        # Clear outdated values from cache
-        self._cache = {}
-
-    @property
-    def lms(self):
-        if self.lmsa:
-            return self.lmsa[:-1]
-
-    @lms.setter
-    def lms(self, color):
-        self.lmsa = color
-
 
 class ColorMixin(object):
     """Mixin class for visual stim that need color and or contrast.
@@ -1317,104 +455,6 @@ class ColorMixin(object):
 
         return desiredRGB
 
-
-class ColorArray(object):
-    def __init__(self, array, space, handler='basic', conematrix=None):
-        self._franca = []
-        self._cache = {}
-        self.set(array, space, handler='basic', conematrix=None)
-
-
-    def set(self, array, space, handler='basic', conematrix=None):
-        # Validate
-        if isinstance(array, numpy.ndarray):
-            # If numpy, convert to list of tuple
-            new = [[[]]*len(array[0])]*len(array)
-            for row in range(len(array)):
-                for col in range(len(array[row])):
-                    new[row][col] = tuple(float(c) for c in array[row][col])
-                    if space in ['rgb255', 'rgba255']:
-                        new[row][col] = tuple(int(c) for c in array[row][col])
-            array = new
-        if not isinstance(array, list):
-            return
-
-        if handler in ['color', 'basic', Color]:
-            self.handler = Color()
-        elif handler in ['advanced', AdvancedColor]:
-            self.handler = AdvancedColor(conematrix=conematrix)
-        elif isinstance(handler, Color):
-            self.handler = handler
-        else:
-            return
-        # If 1d
-        if not isinstance(array[0], list):
-            array = [array]
-        self._franca = [[[]]*len(array[0])]*len(array)
-        for row in range(len(array)):
-            for col in range(len(array[row])):
-                self.handler.set(array[row][col], space)
-                self._franca[row][col] = self.handler.rgba
-        self._cache = {}
-
-    def get(self, space):
-        out = [[[]]*len(self._franca[0])]*len(self._franca)
-        for row in range(len(self._franca)):
-            for col in range(len(self._franca[row])):
-                self.handler.set(self._franca[row][col], 'rgba')
-                out[row][col] = getattr(self.handler, space)
-        return out
-
-    @property
-    def rgba(self):
-        if 'rgba' not in self._cache:
-            self._cache['rgba'] = self.get('rgba')
-        return self._cache['rgba']
-    @property
-    def rgb(self):
-        if 'rgb' not in self._cache:
-            self._cache['rgb'] = self.get('rgb')
-        return self._cache['rgb']
-    @property
-    def rgba1(self):
-        if 'rgba1' not in self._cache:
-            self._cache['rgba1'] = self.get('rgba1')
-        return self._cache['rgba1']
-    @property
-    def rgb1(self):
-        if 'rgb1' not in self._cache:
-            self._cache['rgb1'] = self.get('rgb1')
-        return self._cache['rgb1']
-    @property
-    def rgba255(self):
-        if 'rgba255' not in self._cache:
-            self._cache['rgba255'] = self.get('rgba255')
-        return self._cache['rgba255']
-    @property
-    def rgb255(self):
-        if 'rgb255' not in self._cache:
-            self._cache['rgb255'] = self.get('rgb255')
-        return self._cache['rgb255']
-    @property
-    def hsva(self):
-        if 'hsva' not in self._cache:
-            self._cache['hsva'] = self.get('hsva')
-        return self._cache['hsva']
-    @property
-    def hsv(self):
-        if 'hsv' not in self._cache:
-            self._cache['hsv'] = self.get('hsv')
-        return self._cache['hsv']
-    @property
-    def hexa(self):
-        if 'hexa' not in self._cache:
-            self._cache['hexa'] = self.get('hexa')
-        return self._cache['hexa']
-    @property
-    def hex(self):
-        if 'hex' not in self._cache:
-            self._cache['hex'] = self.get('hex')
-        return self._cache['hex']
 
 class ContainerMixin(object):
     """Mixin class for visual stim that have verticesPix attrib
